% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{acs}
\alias{acs}
\title{Make age group using American Community Survey style}
\usage{
acs()
}
\value{
Creates a breaksStrategy object holding the labels for categorization
}
\description{
Helper function for age characteristic to make age breaks following the break
style of the American Community Survey (acs)
}
