% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addChar.R
\name{addAgeChar}
\alias{addAgeChar}
\title{Add an age characteristic}
\usage{
addAgeChar(clinChar, categorize = NULL)
}
\arguments{
\item{clinChar}{a clinChar object maintaining the components of the characterization}

\item{categorize}{describes how the continuous value should be categorized.
This function takes a breaksStrategy object to describe the categories ow it is left NULL.
If the parameter is NULL then no categoization summary is done}
}
\value{
adds a ageChar object into the clinChar extractSettings slot
}
\description{
This function adds an age characteristic to the clinChar object.
}
