% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{addCohortCovariates}
\alias{addCohortCovariates}
\title{Make cohort settings}
\usage{
addCohortCovariates(settings, cohortId, cohortName, timeA, timeB)
}
\arguments{
\item{settings}{the clinicalCharacteristicsSettings object to add to}

\item{cohortId}{the ids of the cohorts in the initialized cohort table to use as covariates}

\item{cohortName}{the names of the cohorts corresponding to their id}

\item{timeA}{a vector of time points. This should be the left point in the time interval}

\item{timeB}{a vector of time points. This should be the right point in the time interval}
}
\value{
list of settings with cohort covariates
}
\description{
Make cohort settings
}
