% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addChar.R
\name{addConditionCount}
\alias{addConditionCount}
\title{Add a condition count characteristic}
\usage{
addConditionCount(
  clinChar,
  timeWindows,
  conceptSets = NULL,
  categorize = NULL,
  conceptType = NULL,
  sourceConcepts = NULL
)
}
\arguments{
\item{clinChar}{a clinChar object maintaining the components of the characterization}

\item{timeWindows}{a timeWindow object that specifies the boundaries relative to the target start date
on when to search for the presence of a value. use \code{makeTimeTable} function}

\item{conceptSets}{a list of concept sets that specify the codes to search within the domain}

\item{categorize}{describes how the continuous value should be categorized.
This function takes a breaksStrategy object to describe the categories ow it is left NULL.
If the parameter is NULL then no categorization summary is done}

\item{conceptType}{the condition type concept ids to use to limit the count}

\item{sourceConcepts}{the source concept ids to use to limit the count}
}
\value{
adds a countChar object of condition into the clinChar extractSettings slot
}
\description{
This function adds a count characteristic to the clinChar object for a condition.
A count characteristics enumerates the number of events observed during a window of time.
}
