% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{addDefaultEthnicityLineItems}
\alias{addDefaultEthnicityLineItems}
\title{Convenience function to add default ethnicity line items}
\usage{
addDefaultEthnicityLineItems()
}
\value{
a list of line items for default ethnicity categories (hispanic, not hispanic, not reported)
}
\description{
Convenience function to add default ethnicity line items
}
