% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{addDefaultGenderLineItems}
\alias{addDefaultGenderLineItems}
\title{Convenience function to add male and female line items for demographic characterization}
\usage{
addDefaultGenderLineItems()
}
\value{
a list of two line items for male and female gender
}
\description{
Convenience function to add male and female line items for demographic characterization
}
