% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{addDefaultRaceLineItems}
\alias{addDefaultRaceLineItems}
\title{Convenience function to add default race line items}
\usage{
addDefaultRaceLineItems()
}
\value{
a list of line items for default race categories (white, black, asian, not reported)
}
\description{
Convenience function to add default race line items
}
