% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisSettings.R
\name{addDemographicCovariates}
\alias{addDemographicCovariates}
\title{Add demographics settings}
\usage{
addDemographicCovariates(settings, type = c("Categorical", "Continuous"))
}
\arguments{
\item{settings}{the clinicalCharacteristicsSettings object to add to}

\item{type}{the type of demographic covariates to produce, default Categorical and Continuous}
}
\value{
list of settings
}
\description{
Add demographics settings
}
