% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addChar.R
\name{addEthnicityChar}
\alias{addEthnicityChar}
\title{Add a ethnicity characteristic}
\usage{
addEthnicityChar(clinChar)
}
\arguments{
\item{clinChar}{a clinChar object maintaining the components of the characterization}
}
\value{
adds a demoConceptChar object into the clinChar extractSettings slot
}
\description{
This function adds a ethnicity characteristic to the clinChar object.
}
