% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{addLabValueCovariates}
\alias{addLabValueCovariates}
\title{Make lab value covariate settings}
\usage{
addLabValueCovariates(
  settings,
  timeA,
  timeB,
  includeConcepts = c(),
  excludeConcepts = c()
)
}
\arguments{
\item{settings}{the clinicalCharacteristicsSettings object to add to}

\item{timeA}{a vector of time points. This should be the left point in the time interval}

\item{timeB}{a vector of time points. This should be the right point in the time interval}

\item{includeConcepts}{a vector of concept ids limiting the output of what to include}

\item{excludeConcepts}{a vector of concept ids limiting the output of what to exclude}
}
\value{
list of settings adding in lab values
}
\description{
Make lab value covariate settings
}
