% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{addLineItems}
\alias{addLineItems}
\title{Add one or more Line Items to a Section object}
\usage{
addLineItems(section, lineItem, ...)
}
\arguments{
\item{section}{The Section object}

\item{lineItem}{A LineItem object}

\item{...}{Additional LineItem objects to be added}
}
\value{
A Section object
}
\description{
Add one or more Line Items to a Section object
}
