% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addChar.R
\name{addMeasurementCount}
\alias{addMeasurementCount}
\title{Add a measurement count characteristic}
\usage{
addMeasurementCount(
  clinChar,
  timeWindows,
  conceptSets = NULL,
  categorize = NULL,
  conceptType = c(32810, 32869)
)
}
\arguments{
\item{clinChar}{a clinChar object maintaining the components of the characterization}

\item{timeWindows}{a timeWindow object that specifies the boundaries relative to the target start date
on when to search for the presence of a value. use \code{makeTimeTable} function}

\item{conceptSets}{a list of concept sets that specify the codes to search within the domain}

\item{categorize}{describes how the continuous value should be categorized.
This function takes a breaksStrategy object to describe the categories ow it is left NULL.
If the parameter is NULL then no categorization summary is done}
}
\value{
adds a countChar object of measurement  into the clinChar extractSettings slot
}
\description{
This function adds a count characteristic to the clinChar object for a measurement .
A count characteristics enumerates the number of events observed during a window of time.
}
