% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addChar.R
\name{addObservationPresence}
\alias{addObservationPresence}
\title{Add a observation presence characteristic}
\usage{
addObservationPresence(
  clinChar,
  conceptSets,
  timeWindows,
  limit = c("first", "last", "all"),
  conceptType = NULL,
  score = NULL,
  sourceConcepts = NULL
)
}
\arguments{
\item{clinChar}{a clinChar object maintaining the components of the characterization}

\item{conceptSets}{a list of concept sets that specify the codes to search within the domain}

\item{timeWindows}{a timeWindow object that specifies the boundaries relative to the target start date
on when to search for the presence of a value. use \code{makeTimeTable} function}

\item{limit}{specify which values to use in the characteristic. The last variable will pull the last value in the
time window, the first variable will pull the first value in the time window and the
all vairable will pull all values in the time window}

\item{conceptType}{the type concept ids to use to limit the query}

\item{score}{describes how the categorical value should be converted to a continuous score.
This function takes a scoreStrategy object to describe the scoring ow it is left NULL.
If the parameter is NULL then no continuous summary is done}

\item{sourceConcepts}{the source concept ids to use to limit the query}
}
\value{
adds a presenceChar object of observation into the clinChar extractSettings slot
}
\description{
This function adds a presence characteristic to the clinChar object for a observation.
A presence characteristic summarizes whether a person had the event of
interest as described by a set of codes during a window of time.
We use an CIRCE concept set to specify the set of codes to use to determine the presence of an event
in a domain table.
}
