% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{addScoreCovariates}
\alias{addScoreCovariates}
\title{Make score settings}
\usage{
addScoreCovariates(settings, scores = c("CharlsonIndex", "Dcsi", "Chads2Vasc"))
}
\arguments{
\item{settings}{the clinicalCharacteristicsSettings object to add to}

\item{scores}{the score covariates to produce, default CharlsonIndex, DCsi, and Chads2Vasc}
}
\value{
list of settings
}
\description{
Make score settings
}
