% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{addSections}
\alias{addSections}
\title{Add one or more Sections to a TableShell object}
\usage{
addSections(tableShell, section, ...)
}
\arguments{
\item{tableShell}{The TableShell object}

\item{section}{A Section object}

\item{...}{Additional Section objects to be added}
}
\value{
A TableShell object
}
\description{
Add one or more Sections to a TableShell object
}
