% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addChar.R
\name{addSpecialtyChar}
\alias{addSpecialtyChar}
\title{Add a specialty characteristic}
\usage{
addSpecialtyChar(clinChar, visitDetailTable, timeWindows, asCount)
}
\arguments{
\item{clinChar}{a clinChar object maintaining the components of the characterization}

\item{visitDetailTable}{a visitDetailTable object set using the the makeLocTable fn}

\item{timeWindows}{a timeWindow object that specifies the boundaries relative to the target start date
on when to search for the presence of a value. use \code{makeTimeTable} function}

\item{asCount}{a toggle to determine whether to summarize as a count or presence}
}
\value{
adds a visitDetailChar object into the clinChar extractSettings slot
}
\description{
This function adds alocation characteristic to the clinChar object.
}
