% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{addTargetCohorts}
\alias{addTargetCohorts}
\title{Add TargetCohorts to a TableShell object}
\usage{
addTargetCohorts(tableShell, targetCohorts)
}
\arguments{
\item{tableShell}{The TableShell object}

\item{targetCohorts}{A list of TargetCohort objects}
}
\value{
A TableShell object
}
\description{
Add TargetCohorts to a TableShell object
}
