% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{addTargetCohortsFromCsv}
\alias{addTargetCohortsFromCsv}
\title{Add target cohorts from a CSV file}
\usage{
addTargetCohortsFromCsv(tableShell, file)
}
\arguments{
\item{tableShell}{The table shell object to which the target cohorts will be added.}

\item{file}{The path to the CSV file containing the target cohorts data.}
}
\value{
The modified table shell object.
}
\description{
Add target cohorts from a CSV file
}
