% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{addTargetCohortsFromDf}
\alias{addTargetCohortsFromDf}
\title{Add target cohorts from a data frame}
\usage{
addTargetCohortsFromDf(tableShell, df)
}
\arguments{
\item{tableShell}{The table shell object to which the target cohorts will be added.}

\item{df}{The data frame containing the information for the target cohorts.}
}
\value{
The updated table shell object with the target cohorts added.
}
\description{
Add target cohorts from a data frame
}
