% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addChar.R
\name{addTimeInInpatient}
\alias{addTimeInInpatient}
\title{Add a time in inpatient characteristic}
\usage{
addTimeInInpatient(clinChar, categorize = NULL)
}
\arguments{
\item{clinChar}{a clinChar object maintaining the components of the characterization}

\item{categorize}{describes how the continuous value should be categorized.
This function takes a breaksStrategy object to describe the categories ow it is left NULL.
If the parameter is NULL then no categorization summary is done}
}
\value{
adds a timeInChar object into the clinChar extractSettings slot
}
\description{
This function finds the time spent in inpatient hospitalization per patient. This takes
the difference between the visit start date and end date
}
