% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addChar.R
\name{addTimeToCondition}
\alias{addTimeToCondition}
\title{Add a condition timeTo characteristic}
\usage{
addTimeToCondition(
  clinChar,
  conceptSets,
  timeWindows,
  limit = c("first", "last", "all"),
  conceptType = NULL,
  categorize = NULL,
  sourceConcepts = NULL
)
}
\arguments{
\item{clinChar}{a clinChar object maintaining the components of the characterization}

\item{conceptSets}{a list of concept sets that specify the codes to search within the domain}

\item{timeWindows}{a timeWindow object that specifies the boundaries relative to the target start date
on when to search for the presence of a value. use \code{makeTimeTable} function}

\item{limit}{specify which values to use in the characteristic. The last variable will pull the last value in the
time window, the first variable will pull the first value in the time window and the
all vairable will pull all values in the time window}

\item{conceptType}{the type concept ids to use to limit the query}

\item{categorize}{describes how the continuous value should be categorized.
This function takes a breaksStrategy object to describe the categories ow it is left NULL.
If the parameter is NULL then no categorization summary is done}

\item{sourceConcepts}{the source concept ids to use to limit the query}
}
\value{
adds a timeToChar object of condition into the clinChar extractSettings slot
}
\description{
This function adds a timeTo characteristic to the clinChar object for a condition.
A timeTo characteristic summarizes the time to an event of
interest as described by a set of codes during a window of time.
We use an CIRCE concept set to specify the set of codes to use to determine the presence of an event
in a domain table.
}
