% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisSettings.R
\name{addVisitCovariates}
\alias{addVisitCovariates}
\title{Make visit covariate settings}
\usage{
addVisitCovariates(
  settings,
  timeA,
  timeB,
  includeConcepts = c(),
  excludeConcepts = c()
)
}
\arguments{
\item{settings}{the clinicalCharacteristicsSettings object to add to}

\item{timeA}{a vector of time points. This should be the left point in the time interval}

\item{timeB}{a vector of time points. This should be the right point in the time interval}

\item{includeConcepts}{a vector of concept ids limiting the output of what to include}

\item{excludeConcepts}{a vector of concept ids limiting the output of what to exclude}
}
\value{
list of settings adding in visits
}
\description{
Make visit covariate settings
}
