% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addChar.R
\name{addYearChar}
\alias{addYearChar}
\title{Add a year characteristic}
\usage{
addYearChar(clinChar, categorize = NULL)
}
\arguments{
\item{clinChar}{a clinChar object maintaining the components of the characterization}
}
\value{
adds a yearChar object into the clinChar extractSettings slot
}
\description{
This function adds a year characteristic to the clinChar object.
}
