% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{adherentPresenceStat}
\alias{adherentPresenceStat}
\title{Adherent Presence Stat}
\usage{
adherentPresenceStat()
}
\value{
A presence stat object
}
\description{
Create a presence stat where only occurrence during the observation period are valid and the denominator are those
who only adhere to the observation period
}
