% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{age10yrGrp}
\alias{age10yrGrp}
\title{Create a breaks Strategy object for age into 10 year groups}
\usage{
age10yrGrp()
}
\value{
A BreaksStreategy object with defaults assumptions for 10 year age groups
}
\description{
Create a breaks Strategy object for age into 10 year groups
}
