% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{age18}
\alias{age18}
\title{Make age group using 18 threshold}
\usage{
age18()
}
\value{
Creates a breaksStrategy object holding the labels for categorization
}
\description{
Helper function for age characteristic to make age breaks categorizing persons
below age 18 and at or above age 18
}
