% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{age19Grps}
\alias{age19Grps}
\title{Make 19 age groups}
\usage{
age19Grps()
}
\value{
Creates a breaksStrategy object holding the labels for categorization
}
\description{
Helper function for age characteristic to make 19 age groups where
every 5 years are categorized up to age 85. 85+ is a single group
}
