% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{age5yrGrp}
\alias{age5yrGrp}
\title{Create a breaks Strategy object for age into 5 year groups}
\usage{
age5yrGrp()
}
\value{
A BreaksStrategy object with defaults assumptions for 5 year age groups
}
\description{
Create a breaks Strategy object for age into 5 year groups
}
