% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{age65}
\alias{age65}
\title{Make age group using 65 threshold}
\usage{
age65()
}
\value{
Creates a breaksStrategy object holding the labels for categorization
}
\description{
Helper function for age characteristic to make age breaks categorizing persons
below age 65 and at or above age 65
}
