% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{ageAcs}
\alias{ageAcs}
\title{Make age group using American Community Survey style}
\usage{
ageAcs()
}
\value{
Creates a breaksStrategy object holding the labels for categorization
}
\description{
Helper function for age characteristic to make age breaks following the break
style of the American Community Survey (acs)
}
