% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{ageChar}
\alias{ageChar}
\title{Create a age statistic}
\usage{
ageChar(breaks = NULL)
}
\arguments{
\item{breaks}{a breaksStrategy object dictating how to classify counts into categories}
}
\value{
A DemographicAge Statistic class object
}
\description{
Create a age statistic
}
