% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{ageCharCts}
\alias{ageCharCts}
\title{Create a age statistic as continuous}
\usage{
ageCharCts()
}
\value{
A DemographicAge Statistic class object as continuous
}
\description{
Create a age statistic as continuous
}
