% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{ageChildWorkRetire}
\alias{ageChildWorkRetire}
\title{Make age groups for common workforce lifespan}
\usage{
ageChildWorkRetire()
}
\value{
Creates a breaksStrategy object holding the labels for categorization
}
\description{
Helper function for age characteristic to make age breaks following lifespan
of a working human. 0-14 is child, 15-64 is working age, 65+ is retirement
}
