% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{anyCountBreaksStat}
\alias{anyCountBreaksStat}
\title{Any Count with Breaks}
\usage{
anyCountBreaksStat(breaks)
}
\arguments{
\item{breaks}{a breaksStrategy object dictating how to classify counts into categories.
If null then this defaults to a continuous distribution}
}
\value{
A stat object breaks
}
\description{
Create a count stat with breaks where any occurrence is valid.
}
