% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{anyCountCtsStat}
\alias{anyCountCtsStat}
\title{Any Count Continuous}
\usage{
anyCountCtsStat()
}
\value{
A stat object continuousDistribution
}
\description{
Create a count stat where any occurrence is valid.
}
