% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{anyTimeToFirstBreaksStat}
\alias{anyTimeToFirstBreaksStat}
\title{Create a time to stat with breaks where any occurrence is valid}
\usage{
anyTimeToFirstBreaksStat(breaks)
}
\arguments{
\item{breaks}{a breaksStrategy object dictating how to classify counts into categories.
If null then this defaults to a continuous distribution}
}
\value{
A stat object breaks
}
\description{
Create a time to stat with breaks where any occurrence is valid
}
