% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bindCovariateFilesByTime}
\alias{bindCovariateFilesByTime}
\title{Bind Covariate Files by time window}
\usage{
bindCovariateFilesByTime(folder, domain, delete = TRUE)
}
\arguments{
\item{folder}{the folder with the clinical characteristic output}

\item{domain}{the domain to bind, use lead domains with multiple time windows}

\item{delete}{option specifying whether to delete the individual files}
}
\value{
invisible return, saves bound files
}
\description{
Bind Covariate Files by time window
}
