% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{Breaks}
\alias{Breaks}
\title{Breaks Statistic}
\description{
A statistic that converts a continuous value to a categorical value by grouping
the number of events into discrete buckets.
}
\section{Super class}{
\code{\link[ClinicalCharacteristics:Statistic]{ClinicalCharacteristics::Statistic}} -> \code{Breaks}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Breaks-new}{\code{Breaks$new()}}
\item \href{#method-Breaks-clone}{\code{Breaks$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getAggregationType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getAggregationType'><code>ClinicalCharacteristics::Statistic$getAggregationType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getBreaksIfAny"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getBreaksIfAny'><code>ClinicalCharacteristics::Statistic$getBreaksIfAny()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getPersonLineTransformation"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getPersonLineTransformation'><code>ClinicalCharacteristics::Statistic$getPersonLineTransformation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getStatisticType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getStatisticType'><code>ClinicalCharacteristics::Statistic$getStatisticType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getWeightsIfAny"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getWeightsIfAny'><code>ClinicalCharacteristics::Statistic$getWeightsIfAny()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Breaks-new"></a>}}
\if{latex}{\out{\hypertarget{method-Breaks-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Breaks$new(personLine, breaks)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{personLine}}{the means of converting occurrences to a single event per patient.
For presence this could be any, observed or adherent}

\item{\code{breaks}}{a breaks strategy object to categorize results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Breaks-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Breaks-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Breaks$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
