% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{BreaksStrategy}
\alias{BreaksStrategy}
\title{BreaksStrategy}
\description{
An R6 class to define a BreaksStrategy object
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{the name of the breaks strategy}

\item{\code{type}}{the type of breaks strategy. Could be 'value' or 'concept'}

\item{\code{labels}}{A character vector used to label each break interval}

\item{\code{breaks}}{a vector with cut points}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BreaksStrategy-new}{\code{BreaksStrategy$new()}}
\item \href{#method-BreaksStrategy-makeCaseWhenSql}{\code{BreaksStrategy$makeCaseWhenSql()}}
\item \href{#method-BreaksStrategy-clone}{\code{BreaksStrategy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BreaksStrategy-new"></a>}}
\if{latex}{\out{\hypertarget{method-BreaksStrategy-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BreaksStrategy$new(name, labels, breaks, type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{the name of the breaks strategy}

\item{\code{labels}}{a character vector indicating how to label each break interval}

\item{\code{breaks}}{a vector with cut points}

\item{\code{type}}{the type of breaks strategy. Could be 'value' or 'concept'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BreaksStrategy-makeCaseWhenSql"></a>}}
\if{latex}{\out{\hypertarget{method-BreaksStrategy-makeCaseWhenSql}{}}}
\subsection{Method \code{makeCaseWhenSql()}}{
Generate SQL code for a CASE WHEN statement based on the break strategy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BreaksStrategy$makeCaseWhenSql(ordinalId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ordinalId}}{the order identifier of the line item in the table shell}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BreaksStrategy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BreaksStrategy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BreaksStrategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
