% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{BuildOptions}
\alias{BuildOptions}
\title{BuildOptions}
\description{
An R6 class to define build options for the tableShell
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{codesetTempTable}}{table name for codeset table}

\item{\code{sourceCodesetTempTable}}{table name for source codeset table}

\item{\code{timeWindowTempTable}}{table name for time windows}

\item{\code{targetCohortTempTable}}{table name for target cohorts}

\item{\code{tsMetaTempTable}}{table name for table shell meta}

\item{\code{conceptSetOccurrenceTempTable}}{table name for concept set occurrence table}

\item{\code{cohortOccurrenceTempTable}}{table name for cohort occurrence table}

\item{\code{patientLevelDataTempTable}}{table name for patient level data}

\item{\code{patientLevelTableShellTempTable}}{table name for patient level data table merged with ts meta}

\item{\code{categoricalSummaryTempTable}}{table name for categorical summary table}

\item{\code{continuousSummaryTempTable}}{table name for continuous summary table}

\item{\code{useCohortEra}}{toggle to choose if using cohort era or start date}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BuildOptions-new}{\code{BuildOptions$new()}}
\item \href{#method-BuildOptions-clone}{\code{BuildOptions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BuildOptions-new"></a>}}
\if{latex}{\out{\hypertarget{method-BuildOptions-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BuildOptions$new(
  codesetTempTable = NULL,
  sourceCodesetTempTable = NULL,
  timeWindowTempTable = NULL,
  targetCohortTempTable = NULL,
  tsMetaTempTable = NULL,
  conceptSetOccurrenceTempTable = NULL,
  cohortOccurrenceTempTable = NULL,
  patientLevelDataTempTable = NULL,
  patientLevelTableShellTempTable = NULL,
  categoricalSummaryTempTable = NULL,
  continuousSummaryTempTable = NULL,
  useCohortEra = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{codesetTempTable}}{the name of the codeset table used in execution. Defaults as a temp table #codeset}

\item{\code{sourceCodesetTempTable}}{he name of the source codeset table used in execution}

\item{\code{timeWindowTempTable}}{the name of the time Window table used in execution. Defaults as a temp table #time_windows}

\item{\code{targetCohortTempTable}}{the name of the target cohort table used in execution. Defaults as a temp table #target_cohorts}

\item{\code{tsMetaTempTable}}{the name of the table shell meta table used in execution. Defaults as a temp table #ts_meta}

\item{\code{conceptSetOccurrenceTempTable}}{the name of the concept set occurrence table used in execution. Defaults as a temp table #concept_set_occ}

\item{\code{cohortOccurrenceTempTable}}{the name of the cohort occurrence  table used in execution. Defaults as a temp table #cohort_occ}

\item{\code{patientLevelDataTempTable}}{the name of the patient level data table used in execution. Note this does not contain info of the table shell. Defaults as a temp table #patient_data}

\item{\code{patientLevelTableShellTempTable}}{the name of the patient level data table with additional meta info used in execution. Defaults as a temp table #pat_ts_tab}

\item{\code{categoricalSummaryTempTable}}{the name of the categorical summary table used in execution. Defaults as a temp table #categorical_table}

\item{\code{continuousSummaryTempTable}}{the name of the continuous summary table used in execution. Defaults as a temp table #continuous_table}

\item{\code{useCohortEra}}{a true false toggle specifying if in a cohort Char whether to use the cohort era (TRUE) or just the start date (FALSE)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BuildOptions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BuildOptions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BuildOptions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
