% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{charlsonIndexScore}
\alias{charlsonIndexScore}
\title{Function to make charlson score}
\usage{
charlsonIndexScore(ageId)
}
\arguments{
\item{ageId}{the id of the age slot in the clinChar object}
}
\value{
Creates a scoreStrategy object holding weights for score
}
\description{
Function to make charlson score
}
