% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{cohortFollowupTime}
\alias{cohortFollowupTime}
\title{Create a cohort follow up time characteristic}
\usage{
cohortFollowupTime()
}
\value{
A DemographicCohortTime Statistic class object
}
\description{
Create a cohort follow up time characteristic
}
