% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{CohortInfo}
\alias{CohortInfo}
\title{CohortInfoe}
\description{
An R6 class to define a Cohort Info object.
CohortInfo objects do not maintain any execution settings, just the id and name
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CohortInfo-new}{\code{CohortInfo$new()}}
\item \href{#method-CohortInfo-getId}{\code{CohortInfo$getId()}}
\item \href{#method-CohortInfo-getName}{\code{CohortInfo$getName()}}
\item \href{#method-CohortInfo-cohortDetails}{\code{CohortInfo$cohortDetails()}}
\item \href{#method-CohortInfo-clone}{\code{CohortInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-CohortInfo-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortInfo$new(id, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{the cohort definition id}

\item{\code{name}}{the name of the cohort definition}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortInfo-getId"></a>}}
\if{latex}{\out{\hypertarget{method-CohortInfo-getId}{}}}
\subsection{Method \code{getId()}}{
get the cohort id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortInfo$getId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortInfo-getName"></a>}}
\if{latex}{\out{\hypertarget{method-CohortInfo-getName}{}}}
\subsection{Method \code{getName()}}{
get the cohort name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortInfo$getName()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortInfo-cohortDetails"></a>}}
\if{latex}{\out{\hypertarget{method-CohortInfo-cohortDetails}{}}}
\subsection{Method \code{cohortDetails()}}{
print the cohort details
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortInfo$cohortDetails()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
