% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{CohortLineItem}
\alias{CohortLineItem}
\title{CohortLineItem}
\description{
An R6 class to define a CohortLineItem
}
\section{Super class}{
\code{\link[ClinicalCharacteristics:LineItem]{ClinicalCharacteristics::LineItem}} -> \code{CohortLineItem}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CohortLineItem-new}{\code{CohortLineItem$new()}}
\item \href{#method-CohortLineItem-clone}{\code{CohortLineItem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="LineItem" data-id="getLineItemMeta"><a href='../../ClinicalCharacteristics/html/LineItem.html#method-LineItem-getLineItemMeta'><code>ClinicalCharacteristics::LineItem$getLineItemMeta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="LineItem" data-id="getStatistic"><a href='../../ClinicalCharacteristics/html/LineItem.html#method-LineItem-getStatistic'><code>ClinicalCharacteristics::LineItem$getStatistic()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortLineItem-new"></a>}}
\if{latex}{\out{\hypertarget{method-CohortLineItem-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortLineItem$new(
  sectionLabel,
  domainTable,
  covariateCohort,
  timeInterval,
  statistic
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortLineItem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortLineItem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortLineItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
