% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{ConceptSetGroupLineItem}
\alias{ConceptSetGroupLineItem}
\title{ConceptSetGroupLineItem}
\description{
An R6 class to define a ConceptSetGroupLineItem
}
\section{Super class}{
\code{\link[ClinicalCharacteristics:LineItem]{ClinicalCharacteristics::LineItem}} -> \code{ConceptSetGroupLineItem}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConceptSetGroupLineItem-new}{\code{ConceptSetGroupLineItem$new()}}
\item \href{#method-ConceptSetGroupLineItem-grabConceptSet}{\code{ConceptSetGroupLineItem$grabConceptSet()}}
\item \href{#method-ConceptSetGroupLineItem-clone}{\code{ConceptSetGroupLineItem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="LineItem" data-id="getLineItemMeta"><a href='../../ClinicalCharacteristics/html/LineItem.html#method-LineItem-getLineItemMeta'><code>ClinicalCharacteristics::LineItem$getLineItemMeta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="LineItem" data-id="getStatistic"><a href='../../ClinicalCharacteristics/html/LineItem.html#method-LineItem-getStatistic'><code>ClinicalCharacteristics::LineItem$getStatistic()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConceptSetGroupLineItem-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConceptSetGroupLineItem-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConceptSetGroupLineItem$new(
  sectionLabel,
  groupLabel,
  conceptSets,
  domainTables,
  timeInterval,
  statistic
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sectionLabel}}{a label for the table shell section}

\item{\code{groupLabel}}{a label for the group}

\item{\code{conceptSets}}{a group of concept sets}

\item{\code{domainTables}}{the domain tables in the cdm}

\item{\code{timeInterval}}{a time interval class object to determine the time frame to consider the analytic}

\item{\code{statistic}}{a Statistic Class object used to determine what type of analytic should be done for the line item}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConceptSetGroupLineItem-grabConceptSet"></a>}}
\if{latex}{\out{\hypertarget{method-ConceptSetGroupLineItem-grabConceptSet}{}}}
\subsection{Method \code{grabConceptSet()}}{
retrieve the concept sets
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConceptSetGroupLineItem$grabConceptSet()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConceptSetGroupLineItem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConceptSetGroupLineItem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConceptSetGroupLineItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
