% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{ConceptSetLineItem}
\alias{ConceptSetLineItem}
\title{ConceptSetLineItem}
\description{
An R6 class to define a ConceptSetLineItem
}
\section{Super class}{
\code{\link[ClinicalCharacteristics:LineItem]{ClinicalCharacteristics::LineItem}} -> \code{ConceptSetLineItem}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConceptSetLineItem-new}{\code{ConceptSetLineItem$new()}}
\item \href{#method-ConceptSetLineItem-grabConceptSet}{\code{ConceptSetLineItem$grabConceptSet()}}
\item \href{#method-ConceptSetLineItem-clone}{\code{ConceptSetLineItem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="LineItem" data-id="getLineItemMeta"><a href='../../ClinicalCharacteristics/html/LineItem.html#method-LineItem-getLineItemMeta'><code>ClinicalCharacteristics::LineItem$getLineItemMeta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="LineItem" data-id="getStatistic"><a href='../../ClinicalCharacteristics/html/LineItem.html#method-LineItem-getStatistic'><code>ClinicalCharacteristics::LineItem$getStatistic()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConceptSetLineItem-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConceptSetLineItem-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConceptSetLineItem$new(
  sectionLabel,
  domainTable,
  conceptSet,
  timeInterval,
  statistic
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sectionLabel}}{a label for the table shell section}

\item{\code{domainTable}}{the domain table in the cdm}

\item{\code{conceptSet}}{a concept set class from Capr}

\item{\code{timeInterval}}{a time interval class object to determine the time frame to consider the analytic}

\item{\code{statistic}}{a Statistic Class object used to determine what type of analytic should be done for the line item
grabConceptSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConceptSetLineItem-grabConceptSet"></a>}}
\if{latex}{\out{\hypertarget{method-ConceptSetLineItem-grabConceptSet}{}}}
\subsection{Method \code{grabConceptSet()}}{
helper to pull concept Capr class items
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConceptSetLineItem$grabConceptSet()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConceptSetLineItem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConceptSetLineItem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConceptSetLineItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
