% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convertToLabUnitTable}
\alias{convertToLabUnitTable}
\title{Function to convert table to labUnitTable}
\usage{
convertToLabUnitTable(tbl, noZeroConcept = FALSE, ordLimit = 3)
}
\arguments{
\item{tbl}{a tibble with names measurement_concept_id, measurement_name, unit_concept_id and unit_name.
This table most naturally comes from exploreLabUnits function. But may be made custom but not recommended}

\item{noZeroConcept}{a toggle to remove the 0 unit concept id which stands for no matching unit}

\item{ordLimit}{keep the top 3 combos}
}
\value{
a labUnitTable object used in the labChar characterization
}
\description{
This function is meant to convert the tibble output from exploreLabUnits into
a labUnitTable needed for a labChar
}
