% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{CostTotalClass}
\alias{CostTotalClass}
\title{Total Cost Statistic}
\description{
A Cost Statistic that calculates total cost per patient
}
\section{Super class}{
\code{\link[ClinicalCharacteristics:Statistic]{ClinicalCharacteristics::Statistic}} -> \code{CostTotalClass}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CostTotalClass-new}{\code{CostTotalClass$new()}}
\item \href{#method-CostTotalClass-getCostLabel}{\code{CostTotalClass$getCostLabel()}}
\item \href{#method-CostTotalClass-clone}{\code{CostTotalClass$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getAggregationType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getAggregationType'><code>ClinicalCharacteristics::Statistic$getAggregationType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getBreaksIfAny"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getBreaksIfAny'><code>ClinicalCharacteristics::Statistic$getBreaksIfAny()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getPersonLineTransformation"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getPersonLineTransformation'><code>ClinicalCharacteristics::Statistic$getPersonLineTransformation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getStatisticType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getStatisticType'><code>ClinicalCharacteristics::Statistic$getStatisticType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getWeightsIfAny"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getWeightsIfAny'><code>ClinicalCharacteristics::Statistic$getWeightsIfAny()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CostTotalClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-CostTotalClass-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CostTotalClass$new(costCategory, costLine, conceptColumn, conceptId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{costCategory}}{the category name of the demographic}

\item{\code{costLine}}{the line item name of the demographic concept}

\item{\code{conceptColumn}}{the name of column in the person table to extract demographic concept}

\item{\code{conceptId}}{the concept to search for in the concept column}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CostTotalClass-getCostLabel"></a>}}
\if{latex}{\out{\hypertarget{method-CostTotalClass-getCostLabel}{}}}
\subsection{Method \code{getCostLabel()}}{
create a label for the demographic concept
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CostTotalClass$getCostLabel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CostTotalClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CostTotalClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CostTotalClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
