% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{Count}
\alias{Count}
\title{An R6 class to define a Count object}
\description{
Child of Statistic. The Count statistic is a poisson metric the indicates the number of occurrences of a variable
}
\section{Super class}{
\code{\link[ClinicalCharacteristics:Statistic]{ClinicalCharacteristics::Statistic}} -> \code{Count}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Count-new}{\code{Count$new()}}
\item \href{#method-Count-clone}{\code{Count$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getStatType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getStatType'><code>ClinicalCharacteristics::Statistic$getStatType()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Count-new"></a>}}
\if{latex}{\out{\hypertarget{method-Count-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Count$new(breaks = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Count-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Count-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Count$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
