% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createAgeLineItem}
\alias{createAgeLineItem}
\title{Create age demographic line item}
\usage{
createAgeLineItem(breaks = NULL)
}
\arguments{
\item{breaks}{A breaks object describing how to categorize the continuous value}
}
\value{
A Demographic line item type class object
}
\description{
Create age demographic line item
}
