% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createCohortInfo}
\alias{createCohortInfo}
\title{Create a CohortInfo object for a cohort and set its attributes}
\usage{
createCohortInfo(id, name)
}
\arguments{
\item{id}{The ID of the cohort}

\item{name}{The name of the cohort}
}
\value{
A CohortInfo object
}
\description{
Create a CohortInfo object for a cohort and set its attributes
}
