% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createCohortLineItem}
\alias{createCohortLineItem}
\title{Create a cohort line item and set its attributes}
\usage{
createCohortLineItem(
  sectionLabel = NA_character_,
  covariateCohort,
  cohortTable,
  timeInterval,
  statistic
)
}
\arguments{
\item{timeInterval}{The TimeInterval object used for the line item}

\item{statistic}{The Statistic object to be used to evaluate the line item}

\item{name}{(OPTIONAL) The name of the line item (if not provided, the name will be set to the cohort name from the CohortInfo object)}

\item{cohort}{A CohortInfo object}
}
\value{
A CohortLineItem object
}
\description{
Create a cohort line item and set its attributes
}
