% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createCohortLineItemBatch}
\alias{createCohortLineItemBatch}
\title{Create a batch of cohort line items from a list of CohortInfo objects.}
\usage{
createCohortLineItemBatch(
  sectionLabel,
  covariateCohorts,
  cohortTable,
  statistic,
  timeIntervals
)
}
\arguments{
\item{sectionLabel}{(OPTIONAL) The name of the line item (if not provided, the name will be set to the Capr concept set name)}

\item{covariateCohorts}{A list of CohortInfo objects}

\item{cohortTable}{The name of the table where the cohorts are stored}

\item{statistic}{The Statistic object to be used to evaluate the line item}

\item{timeIntervals}{A list of TimeInterval class objects}
}
\value{
A list of CohortLineItem objects
}
\description{
The name of each line item will be set to the name of its cohort from the CohortInfo object.
}
