% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createConceptSetLineItem}
\alias{createConceptSetLineItem}
\title{Create a concept set line item and set its attributes}
\usage{
createConceptSetLineItem(
  sectionLabel = NA_character_,
  domain,
  conceptSet,
  timeInterval,
  statistic,
  sourceConceptSet = NULL,
  typeConceptIds = c(),
  visitOccurrenceConceptIds = c()
)
}
\arguments{
\item{sectionLabel}{(OPTIONAL) The name of the line item (if not provided, the name will be set to the Capr concept set name)}

\item{domain}{The domain of the concept set (must be one of 'condition_occurrence', 'drug_exposure', 'procedure_occurrence', 'observation', 'measurement', 'device_exposure')}

\item{conceptSet}{The Capr concept set object}

\item{timeInterval}{The Time Interval object used for the line item}

\item{statistic}{The Statistic object to be used to evaluate the line item}

\item{sourceConceptSet}{(OPTIONAL) A Capr concept set of source concept IDs to use to limit the concept set}

\item{typeConceptIds}{(OPTIONAL) A list of type concept IDs to use to limit the concept set}

\item{visitOccurrenceConceptIds}{(OPTIONAL) A list of visit occurrence concept IDs to use to limit the concept set}
}
\value{
A ConceptSetLineItem object
}
\description{
Create a concept set line item and set its attributes
}
