% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createConceptSetLineItemBatch}
\alias{createConceptSetLineItemBatch}
\title{Create a batch of concept set line items from a list of Capr concept sets.}
\usage{
createConceptSetLineItemBatch(
  sectionLabel,
  domain,
  conceptSets,
  timeIntervals,
  statistic,
  typeConceptIds = c(),
  visitOccurrenceConceptIds = c()
)
}
\arguments{
\item{sectionLabel}{(OPTIONAL) The name of the line item (if not provided, the name will be set to the Capr concept set name)}

\item{conceptSets}{A list of concept set Capr objects}

\item{timeIntervals}{A list of TimeInterval class objects}

\item{statistic}{The Statistic object to be used to evaluate the line item}

\item{typeConceptIds}{(OPTIONAL) A list of type concept IDs to use to limit the concept set}

\item{visitOccurrenceConceptIds}{(OPTIONAL) A list of visit occurrence concept IDs to use to limit the concept set}
}
\value{
A list of ConceptSetLineItem objects
}
\description{
The name of each line item will be set to the name of its Capr concept set. All line items will use the same statistic, domain, type concepts, and visit concepts. It is not possible to specify source concept IDs.
}
