% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createCostLineItem}
\alias{createCostLineItem}
\title{Create a cost line item and set its attributes}
\usage{
createCostLineItem(statistic)
}
\arguments{
\item{statistic}{The Statistic object to be used to evaluate the line item}
}
\value{
A CostLineItem object
}
\description{
Create a cost line item and set its attributes
}
