% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createDemographicLineItem}
\alias{createDemographicLineItem}
\title{Create a demographic line item and set its attributes}
\usage{
createDemographicLineItem(statistic)
}
\arguments{
\item{statistic}{The Statistic object to be used to evaluate the line item}
}
\value{
A DemographicLineItem object
}
\description{
Create a demographic line item and set its attributes
}
