% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createEthnicityLineItem}
\alias{createEthnicityLineItem}
\title{Create ethnicity demographic line item}
\usage{
createEthnicityLineItem()
}
\value{
A Demographic line item type class object
}
\description{
Create ethnicity demographic line item
}
