% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createGenderLineItem}
\alias{createGenderLineItem}
\title{Create gender demographic line item}
\usage{
createGenderLineItem()
}
\value{
A Demographic line item type class object
}
\description{
Create gender demographic line item
}
