% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createLineItem}
\alias{createLineItem}
\title{Create a LineItem object and set its attributes}
\usage{
createLineItem(name, ordinal)
}
\arguments{
\item{name}{The name of the LineItem}

\item{ordinal}{The ordinal of the LineItem}

\item{showMissing}{Whether to show missing values in the LineItem}

\item{statisticType}{The statistic type of the LineItem}

\item{limit}{The limit of the LineItem}
}
\value{
A LineItem object
}
\description{
Create a LineItem object and set its attributes
}
