% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createRaceLineItem}
\alias{createRaceLineItem}
\title{Create race demographic line item}
\usage{
createRaceLineItem()
}
\value{
A Demographic line item type class object
}
\description{
Create race demographic line item
}
