% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportIt.R
\name{createReport}
\alias{createReport}
\title{Convert ClinChar options into a report}
\usage{
createReport(clinChar, outputFolder = here::here(), previewReport = TRUE)
}
\arguments{
\item{clinChar}{the clinChar object}

\item{outputFolder}{the location to save the report output}

\item{previewReport}{look at the html rendering of the report, default true}
}
\description{
Convert ClinChar options into a report
}
