% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createSection}
\alias{createSection}
\title{Create a Section object and set its attributes}
\usage{
createSection(name, ordinal, lineItems)
}
\arguments{
\item{ordinal}{The ordinal of the Section}

\item{lineItems}{A list of LineItem objects}

\item{title}{The title of the Section}
}
\value{
A Section object
}
\description{
Create a Section object and set its attributes
}
