% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createSourceConceptSetLineItem}
\alias{createSourceConceptSetLineItem}
\title{Create a source concept set line item and set its attributes}
\usage{
createSourceConceptSetLineItem(
  sectionLabel = NA_character_,
  domain,
  sourceConceptSet,
  timeInterval,
  statistic,
  typeConceptIds = c()
)
}
\arguments{
\item{sectionLabel}{(OPTIONAL) The name of the line item (if not provided, the name will be set to the Capr concept set name)}

\item{domain}{The domain of the concept set (must be one of 'Condition', 'Drug', 'Procedure', 'Observation', 'Measurement', 'Device')}

\item{sourceConceptSet}{A SourceConcept R6 object created using the \code{sourceConceptSet} function}

\item{timeInterval}{The Time Interval object used for the line item}

\item{statistic}{The Statistic object to be used to evaluate the line item}

\item{typeConceptIds}{(OPTIONAL) A list of type concept IDs to use to limit the concept set}
}
\value{
A SourceConceptSetLineItem object
}
\description{
Create a source concept set line item and set its attributes
}
