% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createSourceConceptSetLineItemBatch}
\alias{createSourceConceptSetLineItemBatch}
\title{Create a batch of source concept set line items from a list of SourceConceptSet classes.}
\usage{
createSourceConceptSetLineItemBatch(
  sectionLabel,
  domain,
  sourceConceptSets,
  timeIntervals,
  statistic,
  typeConceptIds = c()
)
}
\arguments{
\item{sectionLabel}{(OPTIONAL) The name of the line item (if not provided, the name will be set to the Capr concept set name)}

\item{domain}{The domain of the concept set (must be one of 'Condition', 'Drug', 'Procedure', 'Observation', 'Measurement', 'Device')}

\item{sourceConceptSets}{A list of SourceConcept R6 object created using the \code{sourceConceptSet} function}

\item{timeIntervals}{A list of TimeInterval class objects}

\item{statistic}{The Statistic object to be used to evaluate the line item}

\item{typeConceptIds}{(OPTIONAL) A list of type concept IDs to use to limit the concept set}
}
\value{
A list of SourceConceptSetLineItem objects
}
\description{
Create a batch of source concept set line items from a list of SourceConceptSet classes.
}
