% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createTargetCohort}
\alias{createTargetCohort}
\title{Create a TargetCohort object and set its attributes}
\usage{
createTargetCohort(id, name)
}
\arguments{
\item{id}{The ID of the TargetCohort}

\item{name}{The name of the TargetCohort}
}
\value{
A TargetCohort object
}
\description{
Create a TargetCohort object and set its attributes
}
