% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createTimeWindows}
\alias{createTimeWindows}
\title{Create a Time Window object}
\usage{
createTimeWindows(...)
}
\arguments{
\item{...}{a list of timeInterval objects}
}
\value{
A time window object
}
\description{
Create a Time Window object
}
