% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{customBreaks}
\alias{customBreaks}
\title{Function to make custom categorical breaks}
\usage{
customBreaks(x, breaks, labels)
}
\arguments{
\item{x}{a sequence of values to categorize}

\item{breaks}{the breaks to the x sequenece..see \code{cut}}

\item{labels}{the labels of the sequence}
}
\value{
Creates a breaksStrategy object holding the labels for categorization
}
\description{
Function to make custom categorical breaks
}
