% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{defaultTableShellBuildOptions}
\alias{defaultTableShellBuildOptions}
\title{Default build options to generate table shell}
\usage{
defaultTableShellBuildOptions(
  codesetTempTable = "#codeset",
  sourceCodesetTempTable = "#source_codeset",
  timeWindowTempTable = "#time_windows",
  targetCohortTempTable = "#target_cohorts",
  tsMetaTempTable = "#ts_meta",
  conceptSetOccurrenceTempTable = "#concept_set_occ",
  cohortOccurrenceTempTable = "#cohort_occ",
  patientLevelDataTempTable = "#patient_data",
  patientLevelTableShellTempTable = "#pat_ts_tab",
  categoricalSummaryTempTable = "#categorical_table",
  continuousSummaryTempTable = "#continuous_table",
  cohortAnalysisType = c("era", "startDate")
)
}
\arguments{
\item{codesetTempTable}{the name of the codeset table used in execution. Defaults as a temp table #codeset}

\item{timeWindowTempTable}{the name of the time Window table used in execution. Defaults as a temp table #time_windows}

\item{targetCohortTempTable}{the name of the target cohort table used in execution. Defaults as a temp table #target_cohorts}

\item{tsMetaTempTable}{the name of the table shell meta table used in execution. Defaults as a temp table #ts_meta}

\item{conceptSetOccurrenceTempTable}{the name of the concept set occurrence table used in execution. Defaults as a temp table #concept_set_occ}

\item{cohortOccurrenceTempTable}{the name of the cohort occurrence  table used in execution. Defaults as a temp table #cohort_occ}

\item{patientLevelDataTempTable}{the name of the patient level data table used in execution. Note this does not contain info of the table shell. Defaults as a temp table #patient_data}

\item{patientLevelTableShellTempTable}{the name of the patient level data table with additional meta info used in execution. Defaults as a temp table #pat_ts_tab}

\item{categoricalSummaryTempTable}{the name of the categorical summary table used in execution. Defaults as a temp table #categorical_table}

\item{continuousSummaryTempTable}{the name of the continuous summary table used in execution. Defaults as a temp table #continuous_table}

\item{useCohortEra}{a true false toggle specifying if in a cohort Char whether to use the cohort era (TRUE) or just the start date (FALSE)}
}
\value{
A BuildOptions object
}
\description{
Default build options to generate table shell
}
