% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{defaultYearGrp}
\alias{defaultYearGrp}
\title{Create a breaks Strategy object for year}
\usage{
defaultYearGrp(startYear = NULL)
}
\arguments{
\item{startYear}{the year to start the year group sequence. By default this is the year 2000}
}
\value{
A BreaksStreategy object with defaults assumptions for 5 year age groups
}
\description{
Create a breaks Strategy object for year
}
