% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisSettings.R
\name{defineClinicalCharacteristics}
\alias{defineClinicalCharacteristics}
\title{Define clinical characteristics settings}
\usage{
defineClinicalCharacteristics(targetCohortIds, ...)
}
\arguments{
\item{targetCohortIds}{the target cohorts to characterize}

\item{...}{a set of settings to define}
}
\value{
a list object containing the analysis settings
}
\description{
Define clinical characteristics settings
}
