% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{DemographicAge}
\alias{DemographicAge}
\title{Demographic Age Statistic}
\description{
A Demographic Statistic that calculates age from the person table
}
\section{Super class}{
\code{\link[ClinicalCharacteristics:Statistic]{ClinicalCharacteristics::Statistic}} -> \code{DemographicAge}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DemographicAge-new}{\code{DemographicAge$new()}}
\item \href{#method-DemographicAge-getDemoLabel}{\code{DemographicAge$getDemoLabel()}}
\item \href{#method-DemographicAge-modifyBreaksLabels}{\code{DemographicAge$modifyBreaksLabels()}}
\item \href{#method-DemographicAge-clone}{\code{DemographicAge$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getAggregationType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getAggregationType'><code>ClinicalCharacteristics::Statistic$getAggregationType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getBreaksIfAny"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getBreaksIfAny'><code>ClinicalCharacteristics::Statistic$getBreaksIfAny()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getPersonLineTransformation"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getPersonLineTransformation'><code>ClinicalCharacteristics::Statistic$getPersonLineTransformation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getStatisticType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getStatisticType'><code>ClinicalCharacteristics::Statistic$getStatisticType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getWeightsIfAny"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getWeightsIfAny'><code>ClinicalCharacteristics::Statistic$getWeightsIfAny()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicAge-new"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicAge-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicAge$new(statType, aggType, demoCategory, breaks = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{statType}}{the type of statistic}

\item{\code{aggType}}{the way the metric is reported either categorical or continuous}

\item{\code{demoCategory}}{the name of the demographic category}

\item{\code{breaks}}{a breaks strategy object to categorize results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicAge-getDemoLabel"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicAge-getDemoLabel}{}}}
\subsection{Method \code{getDemoLabel()}}{
retrieve the demographic label
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicAge$getDemoLabel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicAge-modifyBreaksLabels"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicAge-modifyBreaksLabels}{}}}
\subsection{Method \code{modifyBreaksLabels()}}{
update the breaks labels within the statistics class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicAge$modifyBreaksLabels(newLabels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newLabels}}{a character string of new labels for the breaks}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicAge-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicAge-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicAge$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
