% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{DemographicCohortTime}
\alias{DemographicCohortTime}
\title{Demographic Cohort Time Statistic}
\description{
A Demographic Statistic that calculates the time (in days) in the target cohort
}
\section{Super class}{
\code{\link[ClinicalCharacteristics:Statistic]{ClinicalCharacteristics::Statistic}} -> \code{DemographicCohortTime}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DemographicCohortTime-new}{\code{DemographicCohortTime$new()}}
\item \href{#method-DemographicCohortTime-getDemoLabel}{\code{DemographicCohortTime$getDemoLabel()}}
\item \href{#method-DemographicCohortTime-clone}{\code{DemographicCohortTime$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getAggregationType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getAggregationType'><code>ClinicalCharacteristics::Statistic$getAggregationType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getBreaksIfAny"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getBreaksIfAny'><code>ClinicalCharacteristics::Statistic$getBreaksIfAny()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getPersonLineTransformation"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getPersonLineTransformation'><code>ClinicalCharacteristics::Statistic$getPersonLineTransformation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getStatisticType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getStatisticType'><code>ClinicalCharacteristics::Statistic$getStatisticType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getWeightsIfAny"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getWeightsIfAny'><code>ClinicalCharacteristics::Statistic$getWeightsIfAny()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicCohortTime-new"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicCohortTime-new}{}}}
\subsection{Method \code{new()}}{
initialize cohort time stat
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicCohortTime$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicCohortTime-getDemoLabel"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicCohortTime-getDemoLabel}{}}}
\subsection{Method \code{getDemoLabel()}}{
retrieve the demographic label
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicCohortTime$getDemoLabel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicCohortTime-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicCohortTime-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicCohortTime$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
