% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{DemographicConcept}
\alias{DemographicConcept}
\title{Demographic Concept Statistic}
\description{
A Demographic Statistic that considers concepts in person table
}
\section{Super class}{
\code{\link[ClinicalCharacteristics:Statistic]{ClinicalCharacteristics::Statistic}} -> \code{DemographicConcept}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DemographicConcept-new}{\code{DemographicConcept$new()}}
\item \href{#method-DemographicConcept-getConceptColumn}{\code{DemographicConcept$getConceptColumn()}}
\item \href{#method-DemographicConcept-getDemoLabel}{\code{DemographicConcept$getDemoLabel()}}
\item \href{#method-DemographicConcept-getConceptId}{\code{DemographicConcept$getConceptId()}}
\item \href{#method-DemographicConcept-clone}{\code{DemographicConcept$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getAggregationType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getAggregationType'><code>ClinicalCharacteristics::Statistic$getAggregationType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getBreaksIfAny"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getBreaksIfAny'><code>ClinicalCharacteristics::Statistic$getBreaksIfAny()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getPersonLineTransformation"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getPersonLineTransformation'><code>ClinicalCharacteristics::Statistic$getPersonLineTransformation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getStatisticType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getStatisticType'><code>ClinicalCharacteristics::Statistic$getStatisticType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getWeightsIfAny"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getWeightsIfAny'><code>ClinicalCharacteristics::Statistic$getWeightsIfAny()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicConcept-new"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicConcept-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicConcept$new(demoCategory, demoLine, conceptColumn, conceptId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{demoCategory}}{the cateogry name of the demographic}

\item{\code{demoLine}}{the line item name of the demographic concept}

\item{\code{conceptColumn}}{the name of column in the person table to extract demographic concept}

\item{\code{conceptId}}{the concept to search for in the concept column
getConceptColumn}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicConcept-getConceptColumn"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicConcept-getConceptColumn}{}}}
\subsection{Method \code{getConceptColumn()}}{
retrieve the concept column
getDemoLabel
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicConcept$getConceptColumn()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicConcept-getDemoLabel"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicConcept-getDemoLabel}{}}}
\subsection{Method \code{getDemoLabel()}}{
create a label for the demographic concept
getConceptId
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicConcept$getDemoLabel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicConcept-getConceptId"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicConcept-getConceptId}{}}}
\subsection{Method \code{getConceptId()}}{
retrieve the concept id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicConcept$getConceptId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicConcept-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicConcept-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicConcept$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
