% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{DemographicLineItem}
\alias{DemographicLineItem}
\title{DemographicLineItem}
\description{
An R6 class to handle the ...
}
\section{Super class}{
\code{\link[ClinicalCharacteristics:LineItem]{ClinicalCharacteristics::LineItem}} -> \code{DemographicLineItem}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DemographicLineItem-new}{\code{DemographicLineItem$new()}}
\item \href{#method-DemographicLineItem-clone}{\code{DemographicLineItem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="LineItem" data-id="getLineItemMeta"><a href='../../ClinicalCharacteristics/html/LineItem.html#method-LineItem-getLineItemMeta'><code>ClinicalCharacteristics::LineItem$getLineItemMeta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="LineItem" data-id="getStatistic"><a href='../../ClinicalCharacteristics/html/LineItem.html#method-LineItem-getStatistic'><code>ClinicalCharacteristics::LineItem$getStatistic()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicLineItem-new"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicLineItem-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicLineItem$new(statistic = statistic)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{statistic}}{a Statistic Class object used to determine what type of analytic should be done for the line item}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicLineItem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicLineItem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicLineItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
