% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{DemographicLocation}
\alias{DemographicLocation}
\title{Demographic Location Statistic}
\description{
A Demographic Statistic that retrieves and categorizes the location of the persons in the target cohort
}
\section{Super class}{
\code{\link[ClinicalCharacteristics:Statistic]{ClinicalCharacteristics::Statistic}} -> \code{DemographicLocation}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DemographicLocation-new}{\code{DemographicLocation$new()}}
\item \href{#method-DemographicLocation-getDemoLabel}{\code{DemographicLocation$getDemoLabel()}}
\item \href{#method-DemographicLocation-modifyBreaksLabels}{\code{DemographicLocation$modifyBreaksLabels()}}
\item \href{#method-DemographicLocation-clone}{\code{DemographicLocation$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getAggregationType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getAggregationType'><code>ClinicalCharacteristics::Statistic$getAggregationType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getBreaksIfAny"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getBreaksIfAny'><code>ClinicalCharacteristics::Statistic$getBreaksIfAny()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getPersonLineTransformation"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getPersonLineTransformation'><code>ClinicalCharacteristics::Statistic$getPersonLineTransformation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getStatisticType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getStatisticType'><code>ClinicalCharacteristics::Statistic$getStatisticType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getWeightsIfAny"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getWeightsIfAny'><code>ClinicalCharacteristics::Statistic$getWeightsIfAny()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicLocation-new"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicLocation-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicLocation$new(breaks)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{breaks}}{a breaks strategy object to categorize results
getDemoLabel}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicLocation-getDemoLabel"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicLocation-getDemoLabel}{}}}
\subsection{Method \code{getDemoLabel()}}{
retrieve the demographic label
modifyBreaksLabels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicLocation$getDemoLabel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicLocation-modifyBreaksLabels"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicLocation-modifyBreaksLabels}{}}}
\subsection{Method \code{modifyBreaksLabels()}}{
update the breaks labels within the statistics class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicLocation$modifyBreaksLabels(newLabels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newLabels}}{a character string of new labels for the breaks}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicLocation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicLocation-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicLocation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
