% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{DemographicPayerType}
\alias{DemographicPayerType}
\title{Demographic Payer Statistic}
\description{
A Demographic Statistic that retrieves and categorizes the payer type from the payer plan period table
}
\section{Super class}{
\code{\link[ClinicalCharacteristics:Statistic]{ClinicalCharacteristics::Statistic}} -> \code{DemographicPayerType}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DemographicPayerType-new}{\code{DemographicPayerType$new()}}
\item \href{#method-DemographicPayerType-getDemoLabel}{\code{DemographicPayerType$getDemoLabel()}}
\item \href{#method-DemographicPayerType-modifyBreaksLabels}{\code{DemographicPayerType$modifyBreaksLabels()}}
\item \href{#method-DemographicPayerType-clone}{\code{DemographicPayerType$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getAggregationType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getAggregationType'><code>ClinicalCharacteristics::Statistic$getAggregationType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getBreaksIfAny"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getBreaksIfAny'><code>ClinicalCharacteristics::Statistic$getBreaksIfAny()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getPersonLineTransformation"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getPersonLineTransformation'><code>ClinicalCharacteristics::Statistic$getPersonLineTransformation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getStatisticType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getStatisticType'><code>ClinicalCharacteristics::Statistic$getStatisticType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getWeightsIfAny"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getWeightsIfAny'><code>ClinicalCharacteristics::Statistic$getWeightsIfAny()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicPayerType-new"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicPayerType-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicPayerType$new(breaks)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{breaks}}{a breaks strategy object to categorize results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicPayerType-getDemoLabel"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicPayerType-getDemoLabel}{}}}
\subsection{Method \code{getDemoLabel()}}{
retrieve the demographic label
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicPayerType$getDemoLabel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicPayerType-modifyBreaksLabels"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicPayerType-modifyBreaksLabels}{}}}
\subsection{Method \code{modifyBreaksLabels()}}{
update the breaks labels within the statistics class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicPayerType$modifyBreaksLabels(newLabels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newLabels}}{a character string of new labels for the breaks}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicPayerType-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicPayerType-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicPayerType$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
