% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{DemographicRace}
\alias{DemographicRace}
\title{Demographic Race Statistic}
\description{
A Demographic Statistic that retrieves and categorizes the patient race from the person table
}
\section{Super class}{
\code{\link[ClinicalCharacteristics:Statistic]{ClinicalCharacteristics::Statistic}} -> \code{DemographicRace}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DemographicRace-new}{\code{DemographicRace$new()}}
\item \href{#method-DemographicRace-getDemoLabel}{\code{DemographicRace$getDemoLabel()}}
\item \href{#method-DemographicRace-modifyBreaksLabels}{\code{DemographicRace$modifyBreaksLabels()}}
\item \href{#method-DemographicRace-clone}{\code{DemographicRace$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getAggregationType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getAggregationType'><code>ClinicalCharacteristics::Statistic$getAggregationType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getBreaksIfAny"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getBreaksIfAny'><code>ClinicalCharacteristics::Statistic$getBreaksIfAny()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getPersonLineTransformation"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getPersonLineTransformation'><code>ClinicalCharacteristics::Statistic$getPersonLineTransformation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getStatisticType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getStatisticType'><code>ClinicalCharacteristics::Statistic$getStatisticType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getWeightsIfAny"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getWeightsIfAny'><code>ClinicalCharacteristics::Statistic$getWeightsIfAny()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicRace-new"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicRace-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicRace$new(breaks)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{breaks}}{a breaks strategy object to categorize results}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicRace-getDemoLabel"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicRace-getDemoLabel}{}}}
\subsection{Method \code{getDemoLabel()}}{
retrieve the demographic label
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicRace$getDemoLabel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicRace-modifyBreaksLabels"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicRace-modifyBreaksLabels}{}}}
\subsection{Method \code{modifyBreaksLabels()}}{
update the breaks labels within the statistics class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicRace$modifyBreaksLabels(newLabels)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newLabels}}{a character string of new labels for the breaks}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicRace-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicRace-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicRace$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
