% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisSettings.R
\name{demographicSettings}
\alias{demographicSettings}
\title{Make demographics settings}
\usage{
demographicSettings(type = c("Categorical", "Continuous"))
}
\arguments{
\item{type}{the type of demographic covariates to produce, default Categorical and Continuous}
}
\value{
list of settings
}
\description{
Make demographics settings
}
