% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisSettings.R
\name{domainSettings}
\alias{domainSettings}
\title{Make domain settings}
\usage{
domainSettings(
  domain,
  timeA,
  timeB,
  includeConcepts = c(),
  excludeConcepts = c()
)
}
\arguments{
\item{domain}{the concept domain to use. Can be Drugs, Conditions, Procedures, Measurements or Visits}

\item{timeA}{a vector of time points. This should be the left point in the time interval}

\item{timeB}{a vector of time points. This should be the right point in the time interval}

\item{includeConcepts}{a vector of concept ids limiting the output of what to include}

\item{excludeConcepts}{a vector of concept ids limiting the output of what to exclude}
}
\value{
list of settings
}
\description{
Make domain settings
}
