% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{ExecutionSettings}
\alias{ExecutionSettings}
\title{ExecutionSettings}
\description{
An R6 class to define an ExecutionSettings object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExecutionSettings-new}{\code{ExecutionSettings$new()}}
\item \href{#method-ExecutionSettings-getDbms}{\code{ExecutionSettings$getDbms()}}
\item \href{#method-ExecutionSettings-connect}{\code{ExecutionSettings$connect()}}
\item \href{#method-ExecutionSettings-disconnect}{\code{ExecutionSettings$disconnect()}}
\item \href{#method-ExecutionSettings-getConnection}{\code{ExecutionSettings$getConnection()}}
\item \href{#method-ExecutionSettings-clone}{\code{ExecutionSettings$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionSettings-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionSettings-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionSettings$new(
  connectionDetails = NULL,
  connection = NULL,
  cdmDatabaseSchema = NULL,
  workDatabaseSchema = NULL,
  tempEmulationSchema = NULL,
  cohortTable = NULL,
  cdmSourceName = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionSettings-getDbms"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionSettings-getDbms}{}}}
\subsection{Method \code{getDbms()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionSettings$getDbms()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionSettings-connect"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionSettings-connect}{}}}
\subsection{Method \code{connect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionSettings$connect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionSettings-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionSettings-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionSettings$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionSettings-getConnection"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionSettings-getConnection}{}}}
\subsection{Method \code{getConnection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionSettings$getConnection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionSettings-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionSettings-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionSettings$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
