% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{exploreLabUnits}
\alias{exploreLabUnits}
\title{Exploratory function to find common lab-unit combinations}
\usage{
exploreLabUnits(
  connection,
  cdmDatabaseSchema,
  vocabDatabaseSchema = cdmDatabaseSchema,
  labIds
)
}
\arguments{
\item{connection}{a database connection to the OMOP CDM using DatabaseConnector}

\item{cdmDatabaseSchema}{the database schema specifying identifying the cdm tables}

\item{vocabDatabaseSchema}{the database schema specifying identifying the vocabulary tables. By
default we use the same input as the cdmDatabaseSchema}

\item{labIds}{OMOP concept ids of lab measures of interest}
}
\value{
a tibble summarizing the most common lab-unit combinations with their names
}
\description{
This function is meant to be used as a preperatory step to making a labChar slot to clinChar.
The input of the lab characterization is a lab-unit combination to search in the CDM. This
function takes labIds of interest and returns the most common unit ids found in tandem within
the CDM of the database of interest. From this table users can build reasonable lab characterizations
where summaries are based on valid lab-unit combinations. Note many lab measurements do not have
corresponding unit value (noted as unit_concept_id = 0). These values may be useful but require
assumptions about the data to make them usable in an analysis. It is recommended that users stick
to unit values that have a corresponding unit concept.
}
