% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{exportClinicalCharacteristics}
\alias{exportClinicalCharacteristics}
\title{Export the characterization}
\usage{
exportClinicalCharacteristics(tb, saveName, savePath = here::here())
}
\arguments{
\item{tb}{the table created from the tabulateClincalCharacteristics}

\item{saveName}{a labelling name to distinguish the characterization}

\item{savePath}{the folder path to save the csv, defaults to current directory}
}
\value{
saves a continuous and categorical csv output from the characterization
}
\description{
Save the characterization result as a csv in a folder
}
