% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{femaleGender}
\alias{femaleGender}
\title{Create a female concept stat}
\usage{
femaleGender()
}
\value{
A DemographicConcept Statistic class object indicating a female concept
}
\description{
Create a female concept stat
}
