% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptType.R
\name{findConceptType}
\alias{findConceptType}
\title{Function to find the type concept ids within each CDM domain}
\usage{
findConceptType(connection, cdmDatabaseSchema, domain)
}
\arguments{
\item{connection}{the connection to the OMOP data}

\item{cdmDatabaseSchema}{the database schema hosting the OMOP cdm}

\item{domain}{the domain to assess. Options include: condition_occurrence,
drug_exposure, measurement, observation and procedure_occurrence}
}
\value{
this function returns a tibble describing the concept type ids
utilized in a specific OMOP domain
}
\description{
In some databases the type concept id is very important in distinguishing what
type of information is available. A domain can have different provenances such
as information from claims, NLP or EHR. This function helps list all possible
concept types corresponding to an OMOP CDM. You should use this function
to identify the correct concept types for different chars.
}
