% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{generateTableShell}
\alias{generateTableShell}
\title{Function to generate results for the table shell object}
\usage{
generateTableShell(tableShell, executionSettings, buildOptions = NULL)
}
\arguments{
\item{tableShell}{The TableShell object to used for generation}

\item{executionSettings}{The ExecutionSettings object used to generate table shell}

\item{buildOptions}{The BuildOptions object used to generate table shell}
}
\value{
A list containing a tibble for categorical and continuous results
}
\description{
Function to generate results for the table shell object
}
