% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{indexYear}
\alias{indexYear}
\title{Create an index year char}
\usage{
indexYear(breaks = NULL)
}
\arguments{
\item{breaks}{A Breaks object dictating how to classify counts into categories.
If null then this defaults to a continuous distribution}
}
\value{
A DemographicIndexYear Statistic class object
}
\description{
Create an index year char
}
