% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{instantiateExample}
\alias{instantiateExample}
\title{Function to instantiate an example environment}
\usage{
instantiateExample(connectionDetails, executionSettings, outputFolder)
}
\arguments{
\item{connectionDetails}{a list specifying connection information using DatabaseConnector::createConnectionDetails}

\item{executionSettings}{a list specifying the cdmDatabaseSchema, cohortTable, and workDatabaseSchema}

\item{outputFolder}{a path to a folder to save the check sum}
}
\value{
invisble return, builds cohort definitions from package to the dbms
}
\description{
Function to instantiate an example environment
}
