% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{IntervalRate}
\alias{IntervalRate}
\title{Interval Rate Statistic}
\description{
A statistic that calculates the rate of occurrence by taking the number of events per person
in the desired interval and dividing by the observed time during the interval. An interval rate
can either be monthly or yearly.
}
\section{Super class}{
\code{\link[ClinicalCharacteristics:Statistic]{ClinicalCharacteristics::Statistic}} -> \code{IntervalRate}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IntervalRate-new}{\code{IntervalRate$new()}}
\item \href{#method-IntervalRate-clone}{\code{IntervalRate$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getAggregationType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getAggregationType'><code>ClinicalCharacteristics::Statistic$getAggregationType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getBreaksIfAny"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getBreaksIfAny'><code>ClinicalCharacteristics::Statistic$getBreaksIfAny()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getPersonLineTransformation"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getPersonLineTransformation'><code>ClinicalCharacteristics::Statistic$getPersonLineTransformation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getStatisticType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getStatisticType'><code>ClinicalCharacteristics::Statistic$getStatisticType()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getWeightsIfAny"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getWeightsIfAny'><code>ClinicalCharacteristics::Statistic$getWeightsIfAny()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntervalRate-new"></a>}}
\if{latex}{\out{\hypertarget{method-IntervalRate-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalRate$new(interval)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{interval}}{the type of interval to use for the rate. can be either monthly or yearly.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IntervalRate-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IntervalRate-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IntervalRate$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
