% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{LineItem}
\alias{LineItem}
\title{LineItem}
\description{
An R6 class to define a LineItem object.
A LineItem is a single, explicitly defined characterization to appear in a Section.
Derived classes exist off of LineItems
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ordinalId}}{the order identifier of the line item in the table shell}

\item{\code{sectionLabel}}{a label for the table shell section}

\item{\code{lineItemLabel}}{a label for the line item}

\item{\code{valueId}}{the id for the line item; either a codeset id, a concept id or a -999 to indicate no true id}

\item{\code{valueDescription}}{the describer for the value id}

\item{\code{domainTable}}{the domain table in the cdm}

\item{\code{lineItemClass}}{the type of line item (ie Demographic, ConceptSet, SourceConceptSet, ConceptSetGroup, Cohort)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LineItem-new}{\code{LineItem$new()}}
\item \href{#method-LineItem-getLineItemMeta}{\code{LineItem$getLineItemMeta()}}
\item \href{#method-LineItem-getStatistic}{\code{LineItem$getStatistic()}}
\item \href{#method-LineItem-clone}{\code{LineItem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LineItem-new"></a>}}
\if{latex}{\out{\hypertarget{method-LineItem-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LineItem$new(
  sectionLabel,
  lineItemLabel = NA_character_,
  domainTable,
  lineItemClass,
  valueId = NA_integer_,
  valueDescription = NA_integer_,
  statistic,
  timeInterval = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sectionLabel}}{a label for the table shell section}

\item{\code{lineItemLabel}}{a label for the line item}

\item{\code{domainTable}}{the domain table in the cdm}

\item{\code{lineItemClass}}{the type of line item (ie Demographic, ConceptSet, SourceConceptSet, ConceptSetGroup, Cohort)}

\item{\code{valueId}}{the id for the line item; either a codeset id, a concept id or a -999 to indicate no true id}

\item{\code{valueDescription}}{the describer for the value id}

\item{\code{statistic}}{a Statistic Class object used to determine what type of analytic should be done for the line item}

\item{\code{timeInterval}}{a time interval class object to determine the time frame to consider the analytic}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LineItem-getLineItemMeta"></a>}}
\if{latex}{\out{\hypertarget{method-LineItem-getLineItemMeta}{}}}
\subsection{Method \code{getLineItemMeta()}}{
retrieve the line item meta information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LineItem$getLineItemMeta()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LineItem-getStatistic"></a>}}
\if{latex}{\out{\hypertarget{method-LineItem-getStatistic}{}}}
\subsection{Method \code{getStatistic()}}{
retrieve the statistic class object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LineItem$getStatistic()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LineItem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LineItem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LineItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
