% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{lineItems}
\alias{lineItems}
\title{Combine all lineItems to enter into the tableShell slot}
\usage{
lineItems(...)
}
\arguments{
\item{...}{A list of lineItems created from various calls}
}
\value{
a flattened list of lineItems
}
\description{
Combine all lineItems to enter into the tableShell slot
}
