% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{listLabUnitCombos}
\alias{listLabUnitCombos}
\title{Function to list labUnit combos}
\usage{
listLabUnitCombos(labIds, unitList, comboNames = NULL)
}
\arguments{
\item{labIds}{a list of lab ids}

\item{unitList}{a list of unit ids}

\item{comboNames}{a list of combination names}
}
\value{
a list of lab unit combinations
}
\description{
This function lists all possible combinations of labs with units
}
