% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookupSourceConcepts}
\alias{lookupSourceConcepts}
\title{Function to look up source concepts in the OMOP Vocabulary}
\usage{
lookupSourceConcepts(codes, vocabulary, executionSettings)
}
\arguments{
\item{codes}{a character string of codes to search}

\item{vocabulary}{the vocabulary to use in search of codes}

\item{executionSettings}{The ExecutionSettings object used to connect to the dbms}
}
\value{
a tibble of four columns: conceptId, conceptName, conceptCode, vocabularyId
}
\description{
Function to look up source concepts in the OMOP Vocabulary
}
