% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinChar.R
\name{makeClinChar}
\alias{makeClinChar}
\title{Create a clinChar object}
\usage{
makeClinChar(
  targetCohortIds,
  targetCohortNames = NULL,
  dbms,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema = cdmDatabaseSchema,
  workDatabaseSchema,
  cohortTable,
  datTableName = "dat"
)
}
\arguments{
\item{targetCohortIds}{the cohort ids from the cohort table needed for description}

\item{targetCohortNames}{the names of the cohorts corresponding to these ids}

\item{dbms}{the dbms where the OMOP data sits}

\item{cdmDatabaseSchema}{the database schema string specifying where the cdm sits}

\item{vocabularyDatabaseSchema}{the database schema string specifying where the vocabulary sits}

\item{workDatabaseSchema}{the database schema string specifying where the cohor table sits}

\item{cohortTable}{the table where the cohorts are}
}
\value{
makes a clinChar object
}
\description{
This function creates a clinChar object used to run the characterization
}
