% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{makeLabUnitTable}
\alias{makeLabUnitTable}
\title{Function to set lab-unit combo}
\usage{
makeLabUnitTable(connection, vocabDatabaseSchema, labUnitList)
}
\arguments{
\item{connection}{a database connection to the OMOP CDM using DatabaseConnector}

\item{vocabDatabaseSchema}{the database schema specifying identifying the vocabulary tables}

\item{labUnitList}{a list to coerce into a labUnit Table}
}
\value{
a labUnitTable object used in the labChar characterization
}
\description{
This function is meant to be build combinations of lab-unit concepts for the lab characterization
}
