% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{makeLocationTable}
\alias{makeLocationTable}
\title{Make a location table using data from the CDM}
\usage{
makeLocationTable(
  connection,
  cdmDatabaseSchema,
  locationColumn = "location_source_value"
)
}
\arguments{
\item{connection}{a database connection to the OMOP CDM using DatabaseConnector}

\item{cdmDatabaseSchema}{the database schema specifying identifying the cdm tables}

\item{locationColumn}{the column in the location table to draw labels}
}
\value{
a locationTable class object that specifies the locationColumn used and the merge key
of the location Char.
}
\description{
Make a location table using data from the CDM
}
