% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{makeTimeTable}
\alias{makeTimeTable}
\title{Function to make a time table for the time windows slot}
\usage{
makeTimeTable(time_a, time_b)
}
\arguments{
\item{time_a}{the time bound on the left side}

\item{time_b}{the time bound on the right side}
}
\description{
Function to make a time table for the time windows slot
}
