% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{makeVisitDetailTable}
\alias{makeVisitDetailTable}
\title{Function to make visit detail table}
\usage{
makeVisitDetailTable(
  connection,
  vocabDatabaseSchema,
  detailType = c("specialty", "care_site"),
  detailIds
)
}
\arguments{
\item{connection}{a database connection to the OMOP CDM using DatabaseConnector}

\item{vocabDatabaseSchema}{the database schema specifying identifying the vocabulary tables}

\item{detailType}{either specialty or care_site}

\item{detailIds}{the OMOP concept ids to get names for}
}
\value{
a visitDetailTable object used in the visitDetailChar characterization
}
\description{
Function to make visit detail table
}
