% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{maleGender}
\alias{maleGender}
\title{Create a male concept stat}
\usage{
maleGender()
}
\value{
A DemographicConcept Statistic class object indicating a male concept
}
\description{
Create a male concept stat
}
