% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{monthlyRate}
\alias{monthlyRate}
\title{Create a monthly interval rate statistic}
\usage{
monthlyRate()
}
\value{
A stat object of class intervalRate
}
\description{
This statistic sums the number of occurrences of an event in a timeInterval and
divides it by the time (modified by  month) to construct a rate per patient. This
can then be summarized as a continuous variable
}
