% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{newConceptBreaks}
\alias{newConceptBreaks}
\title{Create a breaks Strategy object for categorizing concepts}
\usage{
newConceptBreaks(name, breaks, labels)
}
\arguments{
\item{name}{the name of the breaks}

\item{breaks}{a vector with cut points to user}

\item{labels}{a character vector indicating how to label the cut-point. Can stay NULL where a default label is given}
}
\value{
A BreaksStrategy object of type concept
}
\description{
Create a breaks Strategy object for categorizing concepts
}
