% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{newValueBreaks}
\alias{newValueBreaks}
\title{Create a breaks Strategy object for categorizing}
\usage{
newValueBreaks(name, breaks, labels = NULL)
}
\arguments{
\item{name}{the name of the breaks}

\item{breaks}{A Breaks object dictating how to classify counts into categories.
If null then this defaults to a continuous distribution}

\item{labels}{a character vector indicating how to label the cut-point. Can stay NULL where a default label is given}
}
\value{
A BreaksStreategy object
}
\description{
Create a breaks Strategy object for categorizing
}
