% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{observedCountBreaksStat}
\alias{observedCountBreaksStat}
\title{Observed Count with Breaks}
\usage{
observedCountBreaksStat(breaks)
}
\arguments{
\item{breaks}{a breaksStrategy object dictating how to classify counts into categories.
If null then this defaults to a continuous distribution}
}
\value{
A stat object breaks
}
\description{
Create a count stat with breaks where only occurrence during the observation period are valid
}
