% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{observedCountCtsStat}
\alias{observedCountCtsStat}
\title{Observed Count Continuous}
\usage{
observedCountCtsStat()
}
\value{
A stat object continuousDistribution
}
\description{
Create a count stat where only occurrence during the observation period are valid
}
