% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{observedPresenceStat}
\alias{observedPresenceStat}
\title{Observed Presence Stat}
\usage{
observedPresenceStat()
}
\value{
A presence stat object
}
\description{
Create a presence stat where only occurrence during the observation period are valid
}
