% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{observedTimeToFirst}
\alias{observedTimeToFirst}
\title{Observed Time To First}
\usage{
observedTimeToFirst()
}
\value{
A stat object continuousDistribution
}
\description{
Create a time to stat where only occurrence during the observation period is valid
}
