% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{observedTimeToFirstBreaksStat}
\alias{observedTimeToFirstBreaksStat}
\title{Create a time to stat with breaks where only occurrence during the observation period are valid}
\usage{
observedTimeToFirstBreaksStat(breaks)
}
\arguments{
\item{breaks}{A Breaks object dictating how to classify counts into categories.
If null then this defaults to a continuous distribution}
}
\value{
A stat object breaks
}
\description{
Create a time to stat with breaks where only occurrence during the observation period are valid
}
