% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{parseCohortInfoFromDf}
\alias{parseCohortInfoFromDf}
\title{Parse cohort info from a data frame}
\usage{
parseCohortInfoFromDf(df)
}
\arguments{
\item{df}{The data frame containing the information for the cohorts (id and name)}
}
\value{
A list of CohortInfo objects
}
\description{
Parse cohort info from a data frame
}
