% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{parseTargetCohortsFromDf}
\alias{parseTargetCohortsFromDf}
\title{Parse target cohorts from a data frame}
\usage{
parseTargetCohortsFromDf(df)
}
\arguments{
\item{df}{The data frame containing the information for the target cohorts (id and name)}
}
\value{
A list of TargetCohort objects
}
\description{
Parse target cohorts from a data frame
}
