% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{payerType}
\alias{payerType}
\title{Create an payer type char}
\usage{
payerType(breaks = NULL)
}
\arguments{
\item{breaks}{a breaksStrategy object dictating how to classify payer types into categories. by default this will use the SOPT vocabulary}
}
\value{
A DemographicPayerType Statistic class object
}
\description{
Create an payer type char
}
