% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{personLocation}
\alias{personLocation}
\title{Create a location characteristic}
\usage{
personLocation(breaks)
}
\arguments{
\item{breaks}{a breaksStrategy object dictating how to classify locations into categories.}
}
\value{
A DemographicLocation Statistic class object
}
\description{
Create a location characteristic
}
