% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{previewClincalCharacteristics}
\alias{previewClincalCharacteristics}
\title{Preview the characterization}
\usage{
previewClincalCharacteristics(tb, type = c("categorical", "continuous"))
}
\arguments{
\item{tb}{the table created from the tabulateClincalCharacteristics}

\item{type}{which table to preview}
}
\value{
presents the reactable summarizing the characterization
}
\description{
This shows the characterization tables using reactable
}
