% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{raceCategory}
\alias{raceCategory}
\title{Create a race char}
\usage{
raceCategory(breaks = NULL)
}
\arguments{
\item{breaks}{a breaksStrategy object dictating how to classify race into categories. by default this will use custom race categories}
}
\value{
A DemographicRace Statistic class object
}
\description{
Create a race char
}
