% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stow.R
\name{retrieveCount}
\alias{retrieveCount}
\title{Function that retrieves the count table generated from the clinChar}
\usage{
retrieveCount(clinChar)
}
\arguments{
\item{clinChar}{the clinChar object describing the characterization}
}
\value{
A table providing the person count of the cohorts as an arrow object. To get the table you must
pipe \code{dplyr::collect}
}
\description{
Function that retrieves the count table generated from the clinChar
}
