% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stow.R
\name{retrieveTable}
\alias{retrieveTable}
\title{Function that retrieves the data table generated from the clinChar}
\usage{
retrieveTable(clinChar, category_id = NULL)
}
\arguments{
\item{clinChar}{the clinChar object describing the characterization}

\item{category_id}{the slot number of the char object in the extractSettings}
}
\value{
A subset of the data table as an arrow object. To get the table you must
pipe \code{dplyr::collect}
}
\description{
Function that retrieves the data table generated from the clinChar
}
