% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinChar.R
\name{reviewQuery}
\alias{reviewQuery}
\title{Function to review the query built by clinical characteristics}
\usage{
reviewQuery(clinChar, savePath = here::here("clinCharQuery.sql"))
}
\arguments{
\item{clinChar}{the clinChar object}

\item{savePath}{the location to save the query to review. By default
this saves the query to a file named clinCharQuery.sql in your active
directory}
}
\value{
a monaco html widget to review the query. You may edit the query in the
viewer and save to the file created by this function
}
\description{
Function to review the query built by clinical characteristics
}
