% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinChar.R
\name{runClinicalCharacteristics}
\alias{runClinicalCharacteristics}
\title{Runs the characterization}
\usage{
runClinicalCharacteristics(connection, clinChar, dropDat = TRUE)
}
\arguments{
\item{connection}{the DatabaseConnector connection linking to the dbms with OMOP data}

\item{clinChar}{the clinChar object describing the study}

\item{dropDat}{toggle option to drop temporary data table with clinChar results in the dbms}
}
\value{
list of 2 tibbles for continuous and categorical tables
}
\description{
This runs the characterization specified by the clinChar object. The output of the command
is a list containing a tibble of the continuous and categorical output.
}
