% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinChar.R
\name{saveClinicalCharacteristics}
\alias{saveClinicalCharacteristics}
\title{Saves the characterization}
\usage{
saveClinicalCharacteristics(
  clinCharResults,
  saveName = NULL,
  savePath = here::here()
)
}
\arguments{
\item{clinCharResults}{the output from runClinicalCharacteristics}

\item{saveName}{a labelling name to distinguish the characterization}

\item{savePath}{the folder path to save the csv, defaults to current directory}
}
\value{
runs database query described in extractSettings and uploads them to the stow object
}
\description{
This saves the characterization tables from runClinicalCharacteristics
The output is to 2 files one for continuous and other for categorical
}
