% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{saveTableShellResults}
\alias{saveTableShellResults}
\title{Function that previews sql script used to generate results for table shell}
\usage{
saveTableShellResults(result, saveName, savePath = here::here())
}
\arguments{
\item{result}{the list output from \code{generateTableShell} containing a categorical and continuous tibble}

\item{saveName}{The save name of the csv files}

\item{savePath}{the folder location to save the csv files}
}
\value{
A sql file written to a specific location
}
\description{
Function that previews sql script used to generate results for table shell
}
