% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisSettings.R
\name{scoreSettings}
\alias{scoreSettings}
\title{Make score settings}
\usage{
scoreSettings(scores = c("CharlsonIndex", "Dcsi", "Chads2Vasc"))
}
\arguments{
\item{scores}{the score covariates to produce, default CharlsonIndex, DCsi, and Chads2Vasc}
}
\value{
list of settings
}
\description{
Make score settings
}
