% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{setExecutionSettings}
\alias{setExecutionSettings}
\title{Set the ExecutionSettings of a TableShell object}
\usage{
setExecutionSettings(tableShell, executionSettings)
}
\arguments{
\item{tableShell}{The TableShell object}

\item{executionSettings}{The ExecutionSettings object}
}
\value{
A TableShell object
}
\description{
Set the ExecutionSettings of a TableShell object
}
