% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{sourceConceptSet}
\alias{sourceConceptSet}
\title{Function to create a source concept set}
\usage{
sourceConceptSet(sourceConceptTable, name)
}
\arguments{
\item{sourceConceptTable}{a dataframe with source concepts from the OMOP vocabulary}

\item{name}{the name of source concept set}
}
\value{
a SourceConceptSet R6 class specifing the source concepts in use
}
\description{
Function to create a source concept set
}
