% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{SourceConceptSetLineItem}
\alias{SourceConceptSetLineItem}
\title{SourceConceptSetLineItem}
\description{
An R6 class to define a SourceConceptSetLineItem
}
\section{Super class}{
\code{\link[ClinicalCharacteristics:LineItem]{ClinicalCharacteristics::LineItem}} -> \code{SourceConceptSetLineItem}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SourceConceptSetLineItem-new}{\code{SourceConceptSetLineItem$new()}}
\item \href{#method-SourceConceptSetLineItem-grabSourceConceptSet}{\code{SourceConceptSetLineItem$grabSourceConceptSet()}}
\item \href{#method-SourceConceptSetLineItem-clone}{\code{SourceConceptSetLineItem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="LineItem" data-id="getLineItemMeta"><a href='../../ClinicalCharacteristics/html/LineItem.html#method-LineItem-getLineItemMeta'><code>ClinicalCharacteristics::LineItem$getLineItemMeta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="LineItem" data-id="getStatistic"><a href='../../ClinicalCharacteristics/html/LineItem.html#method-LineItem-getStatistic'><code>ClinicalCharacteristics::LineItem$getStatistic()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SourceConceptSetLineItem-new"></a>}}
\if{latex}{\out{\hypertarget{method-SourceConceptSetLineItem-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SourceConceptSetLineItem$new(
  sectionLabel,
  domainTable,
  sourceConceptSet,
  timeInterval,
  statistic
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sectionLabel}}{a label for the table shell section}

\item{\code{domainTable}}{the domain table in the cdm}

\item{\code{sourceConceptSet}}{a source concept Set}

\item{\code{timeInterval}}{a time interval class object to determine the time frame to consider the analytic}

\item{\code{statistic}}{a Statistic Class object used to determine what type of analytic should be done for the line item
grabSourceConceptSet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SourceConceptSetLineItem-grabSourceConceptSet"></a>}}
\if{latex}{\out{\hypertarget{method-SourceConceptSetLineItem-grabSourceConceptSet}{}}}
\subsection{Method \code{grabSourceConceptSet()}}{
retrieve the source concept set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SourceConceptSetLineItem$grabSourceConceptSet()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SourceConceptSetLineItem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SourceConceptSetLineItem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SourceConceptSetLineItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
