% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{TableShell}
\alias{TableShell}
\title{Table Shell}
\description{
An R6 class to define a TableShell object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TableShell-new}{\code{TableShell$new()}}
\item \href{#method-TableShell-getTitle}{\code{TableShell$getTitle()}}
\item \href{#method-TableShell-getTableShellMeta}{\code{TableShell$getTableShellMeta()}}
\item \href{#method-TableShell-getTargetCohorts}{\code{TableShell$getTargetCohorts()}}
\item \href{#method-TableShell-getLineItems}{\code{TableShell$getLineItems()}}
\item \href{#method-TableShell-printJobDetails}{\code{TableShell$printJobDetails()}}
\item \href{#method-TableShell-buildTableShellSql}{\code{TableShell$buildTableShellSql()}}
\item \href{#method-TableShell-outputResults}{\code{TableShell$outputResults()}}
\item \href{#method-TableShell-dropTempTables}{\code{TableShell$dropTempTables()}}
\item \href{#method-TableShell-clone}{\code{TableShell$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableShell-new"></a>}}
\if{latex}{\out{\hypertarget{method-TableShell-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableShell$new(title, targetCohorts, lineItems)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{the title of the table shell}

\item{\code{targetCohorts}}{a list of CohortInfo class objects that describe the index cohorts}

\item{\code{lineItems}}{a list of line item class objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableShell-getTitle"></a>}}
\if{latex}{\out{\hypertarget{method-TableShell-getTitle}{}}}
\subsection{Method \code{getTitle()}}{
get the title of the table shell
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableShell$getTitle()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableShell-getTableShellMeta"></a>}}
\if{latex}{\out{\hypertarget{method-TableShell-getTableShellMeta}{}}}
\subsection{Method \code{getTableShellMeta()}}{
get the meta information for the table shell build
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableShell$getTableShellMeta()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableShell-getTargetCohorts"></a>}}
\if{latex}{\out{\hypertarget{method-TableShell-getTargetCohorts}{}}}
\subsection{Method \code{getTargetCohorts()}}{
get the target cohorts from the table shell
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableShell$getTargetCohorts()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableShell-getLineItems"></a>}}
\if{latex}{\out{\hypertarget{method-TableShell-getLineItems}{}}}
\subsection{Method \code{getLineItems()}}{
get the lineItems from the table shell
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableShell$getLineItems()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableShell-printJobDetails"></a>}}
\if{latex}{\out{\hypertarget{method-TableShell-printJobDetails}{}}}
\subsection{Method \code{printJobDetails()}}{
print the job details of the table shell
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableShell$printJobDetails()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableShell-buildTableShellSql"></a>}}
\if{latex}{\out{\hypertarget{method-TableShell-buildTableShellSql}{}}}
\subsection{Method \code{buildTableShellSql()}}{
function creates the table shell sql needed for the execution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableShell$buildTableShellSql(executionSettings, buildOptions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{executionSettings}}{an executionSettings class obj}

\item{\code{buildOptions}}{a buildOptions class obj}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableShell-outputResults"></a>}}
\if{latex}{\out{\hypertarget{method-TableShell-outputResults}{}}}
\subsection{Method \code{outputResults()}}{
retrieves results from dbms and formats for review
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableShell$outputResults(executionSettings, buildOptions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{executionSettings}}{an executionSettings class obj}

\item{\code{buildOptions}}{a buildOptions class obj}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableShell-dropTempTables"></a>}}
\if{latex}{\out{\hypertarget{method-TableShell-dropTempTables}{}}}
\subsection{Method \code{dropTempTables()}}{
drop all temp tables from the tableShell build
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableShell$dropTempTables(executionSettings, buildOptions)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{executionSettings}}{an executionSettings class obj}

\item{\code{buildOptions}}{a buildOptions class obj}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TableShell-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TableShell-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TableShell$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
