% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{tabulateClinicalCharacteristics}
\alias{tabulateClinicalCharacteristics}
\title{Summarizes the characterization and pulls into R}
\usage{
tabulateClinicalCharacteristics(clinChar)
}
\arguments{
\item{clinChar}{the clinChar object describing the study}
}
\value{
pulls the aggregated data into R
}
\description{
This summarizes the characterization specified by the clinChar object
}
