% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{TimeInterval}
\alias{TimeInterval}
\title{TimeInterval}
\description{
An R6 class to define a TimeInterval
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimeInterval-new}{\code{TimeInterval$new()}}
\item \href{#method-TimeInterval-getLb}{\code{TimeInterval$getLb()}}
\item \href{#method-TimeInterval-getRb}{\code{TimeInterval$getRb()}}
\item \href{#method-TimeInterval-getTimeLabel}{\code{TimeInterval$getTimeLabel()}}
\item \href{#method-TimeInterval-getTimeInterval}{\code{TimeInterval$getTimeInterval()}}
\item \href{#method-TimeInterval-clone}{\code{TimeInterval$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeInterval-new"></a>}}
\if{latex}{\out{\hypertarget{method-TimeInterval-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeInterval$new(lb, rb)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lb}}{left bound - the start of the time interval}

\item{\code{rb}}{right bound - the end of the time interval}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeInterval-getLb"></a>}}
\if{latex}{\out{\hypertarget{method-TimeInterval-getLb}{}}}
\subsection{Method \code{getLb()}}{
return the left bound
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeInterval$getLb()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeInterval-getRb"></a>}}
\if{latex}{\out{\hypertarget{method-TimeInterval-getRb}{}}}
\subsection{Method \code{getRb()}}{
return the right bound
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeInterval$getRb()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeInterval-getTimeLabel"></a>}}
\if{latex}{\out{\hypertarget{method-TimeInterval-getTimeLabel}{}}}
\subsection{Method \code{getTimeLabel()}}{
create and return time labels for left and right bounds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeInterval$getTimeLabel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeInterval-getTimeInterval"></a>}}
\if{latex}{\out{\hypertarget{method-TimeInterval-getTimeInterval}{}}}
\subsection{Method \code{getTimeInterval()}}{
return a tibble with the left and right bounds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeInterval$getTimeInterval()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeInterval-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeInterval-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeInterval$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
