% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{TimeIntervalClass}
\alias{TimeIntervalClass}
\title{TimeIntervalClass}
\description{
An R6 class to define a TimeIntervalClass
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TimeIntervalClass-new}{\code{TimeIntervalClass$new()}}
\item \href{#method-TimeIntervalClass-getLb}{\code{TimeIntervalClass$getLb()}}
\item \href{#method-TimeIntervalClass-getRb}{\code{TimeIntervalClass$getRb()}}
\item \href{#method-TimeIntervalClass-getTimeLabel}{\code{TimeIntervalClass$getTimeLabel()}}
\item \href{#method-TimeIntervalClass-getTimeInterval}{\code{TimeIntervalClass$getTimeInterval()}}
\item \href{#method-TimeIntervalClass-clone}{\code{TimeIntervalClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeIntervalClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-TimeIntervalClass-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeIntervalClass$new(lb, rb)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lb}}{left bound - the start of the time interval}

\item{\code{rb}}{right bound - the end of the time interval}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeIntervalClass-getLb"></a>}}
\if{latex}{\out{\hypertarget{method-TimeIntervalClass-getLb}{}}}
\subsection{Method \code{getLb()}}{
return the left bound
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeIntervalClass$getLb()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeIntervalClass-getRb"></a>}}
\if{latex}{\out{\hypertarget{method-TimeIntervalClass-getRb}{}}}
\subsection{Method \code{getRb()}}{
return the right bound
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeIntervalClass$getRb()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeIntervalClass-getTimeLabel"></a>}}
\if{latex}{\out{\hypertarget{method-TimeIntervalClass-getTimeLabel}{}}}
\subsection{Method \code{getTimeLabel()}}{
create and return time labels for left and right bounds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeIntervalClass$getTimeLabel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeIntervalClass-getTimeInterval"></a>}}
\if{latex}{\out{\hypertarget{method-TimeIntervalClass-getTimeInterval}{}}}
\subsection{Method \code{getTimeInterval()}}{
return a tibble with the left and right bounds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeIntervalClass$getTimeInterval()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TimeIntervalClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TimeIntervalClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeIntervalClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
