% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{timeToFirst}
\alias{timeToFirst}
\title{Create a time to stat where any occurrence is valid}
\usage{
timeToFirst()
}
\value{
A stat object continuousDistribution
}
\description{
Create a time to stat where any occurrence is valid
}
