% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{timeToFirstStat}
\alias{timeToFirstStat}
\title{Create a time to stat where any occurrence is valid}
\usage{
timeToFirstStat(options = c("continuousDistribution", "breaks"), breaks = NULL)
}
\arguments{
\item{options}{set whether to summarize as continuousDistribution or breaks.
A breaks option summarizes as categorial and requires a breaksStrategy}

\item{breaks}{a breaksStrategy object dictating how to classify counts into categories}
}
\value{
A stat object either continuousDistribution or breaks based on the options input
}
\description{
Create a time to stat where any occurrence is valid
}
