% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{year10yrGrp}
\alias{year10yrGrp}
\title{Make year group using 10 yr threshold}
\usage{
year10yrGrp()
}
\value{
Creates a breaksStrategy object holding the labels for categorization
}
\description{
Helper function for year characteristic to make year breaks categorizing persons
per 10 year groups
}
