% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{year5yrGrp}
\alias{year5yrGrp}
\title{Make year group using 5 yr threshold}
\usage{
year5yrGrp(startYear = 2000)
}
\arguments{
\item{startYear}{the start year of the sequence, defaults to 2000}
}
\value{
Creates a breaksStrategy object holding the labels for categorization
}
\description{
Helper function for year characteristic to make year breaks categorizing persons
per 5 year groups
}
