% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breaks.R
\name{yearCovid}
\alias{yearCovid}
\title{Make year group by covid time}
\usage{
yearCovid()
}
\value{
Creates a breaksStrategy object holding the labels for categorization
}
\description{
Helper function for year characteristic to make year breaks categorizing persons
by covid time where 2000-2019 is pre-covid, 2020-2022 is covid and 2023+ is post-covid
}
