% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{yearlyRate}
\alias{yearlyRate}
\title{Create a yearly interval rate statistic}
\usage{
yearlyRate()
}
\value{
A stat object of class intervalRate
}
\description{
This statistic sums the number of occurrences of an event in a timeInterval and
divides it by the time (modified by  year) to construct a rate per patient. This
can then be summarized as a continuous variable
}
