% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addChar.R
\name{addLocationChar}
\alias{addLocationChar}
\title{Add a location characteristic}
\usage{
addLocationChar(clinChar, locationTable)
}
\arguments{
\item{clinChar}{a clinChar object maintaining the components of the characterization}

\item{locationTable}{a locationTable object set using the the makeLocTable fn}
}
\value{
adds a locationChar object into the clinChar extractSettings slot
}
\description{
This function adds alocation characteristic to the clinChar object.
}
