% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{age5yrGrp}
\alias{age5yrGrp}
\title{Make 5 yr age group}
\usage{
age5yrGrp()
}
\value{
Creates a breaksStrategy object holding the labels for categorization
}
\description{
Helper function for age characteristic to make 5 year age breaks
}
