% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{ageCharBreaks}
\alias{ageCharBreaks}
\title{Create a age statistic with breaks}
\usage{
ageCharBreaks(breaks)
}
\arguments{
\item{breaks}{a breaksStrategy object dictating how to classify counts into categories}
}
\value{
A DemographicAge Statistic class object with breaks
}
\description{
Create a age statistic with breaks
}
