% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createCohortLineItem}
\alias{createCohortLineItem}
\title{Create a cohort line item and set its attributes}
\usage{
createCohortLineItem(name = NULL, statistic, cohort, timeInterval)
}
\arguments{
\item{name}{(OPTIONAL) The name of the line item (if not provided, the name will be set to the cohort name from the CohortInfo object)}

\item{statistic}{The Statistic object to be used to evaluate the line item}

\item{cohort}{A CohortInfo object}

\item{timeInterval}{The TimeInterval object used for the line item}
}
\value{
A CohortLineItem object
}
\description{
Create a cohort line item and set its attributes
}
