% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createCohortLineItemBatch}
\alias{createCohortLineItemBatch}
\title{Create a batch of cohort line items from a list of CohortInfo objects.}
\usage{
createCohortLineItemBatch(
  sectionLabel,
  covariateCohorts,
  cohortTable,
  statistic,
  timeIntervals
)
}
\arguments{
\item{sectionLabel}{The name of the cohort batch}

\item{statistic}{The Statistic object to be used to evaluate the line items}

\item{timeIntervals}{A list of TimeInterval class objects}

\item{cohorts}{A list of CohortInfo objects}
}
\value{
A list of CohortLineItem objects
}
\description{
The name of each line item will be set to the name of its cohort from the CohortInfo object.
}
