% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createConceptSetGroupLineItem}
\alias{createConceptSetGroupLineItem}
\title{Create a concept set group item and set its attributes}
\usage{
createConceptSetGroupLineItem(
  sectionLabel = NA_character_,
  groupLabel,
  conceptSets,
  domainTables,
  timeInterval,
  statistic
)
}
\arguments{
\item{sectionLabel}{(OPTIONAL) The name of the line item (if not provided, the name will be the same as the group label)}

\item{groupLabel}{the label of the group}

\item{conceptSets}{A list of Capr concept set object}

\item{domainTables}{a vector of domains corresponding to the concept set}

\item{timeInterval}{The TimeInterval object used for the line item}

\item{statistic}{The Statistic object to be used to evaluate the line item}
}
\value{
A CohortLineItem object
}
\description{
Create a concept set group item and set its attributes
}
