% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createConceptSetLineItemBatch}
\alias{createConceptSetLineItemBatch}
\title{Create a batch of concept set line items from a list of Capr concept sets.}
\usage{
createConceptSetLineItemBatch(
  sectionLabel,
  domain,
  conceptSets,
  timeIntervals,
  statistic,
  typeConceptIds = c(),
  visitOccurrenceConceptIds = c()
)
}
\arguments{
\item{domain}{The domain of the concept sets (must be one of 'Condition', 'Drug', 'Procedure', 'Observation', 'Measurement', 'Device')}

\item{conceptSets}{A list of concept set Capr objects}

\item{timeIntervals}{A list of TimeInterval class objects}

\item{statistic}{The Statistic object to be used to evaluate the line items}

\item{typeConceptIds}{(OPTIONAL) A list of type concept IDs to use to limit the concept set}

\item{visitOccurrenceConceptIds}{(OPTIONAL) A list of visit occurrence concept IDs to use to limit the concept set}

\item{name}{The name of the concept set batch}
}
\value{
A list of ConceptSetLineItem objects
}
\description{
The name of each line item will be set to the name of its Capr concept set. All line items will use the same statistic, domain, type concepts, and visit concepts. It is not possible to specify source concept IDs.
}
