% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createExecutionSettings}
\alias{createExecutionSettings}
\title{Create an ExecutionSettings object and set its attributes}
\usage{
createExecutionSettings(
  connectionDetails,
  connection = NULL,
  cdmDatabaseSchema,
  workDatabaseSchema,
  tempEmulationSchema,
  targetCohortTable,
  cdmSourceName
)
}
\arguments{
\item{connectionDetails}{A DatabaseConnector connectionDetails object (optional if connection is specified)}

\item{connection}{A DatabaseConnector connection object (optional if connectionDetails is specified)}

\item{cdmDatabaseSchema}{The schema of the OMOP CDM database}

\item{workDatabaseSchema}{The schema to which results will be written}

\item{tempEmulationSchema}{Some database platforms like Oracle and Snowflake do not truly support temp tables. To emulate temp tables, provide a schema with write privileges where temp tables can be created.}

\item{targetCohortTable}{The name of the table where the target cohort(s) are stored}

\item{cdmSourceName}{A human-readable name for the OMOP CDM source}
}
\value{
An ExecutionSettings object
}
\description{
Create an ExecutionSettings object and set its attributes
}
