% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createSection}
\alias{createSection}
\title{Create a Section object and set its title and ordinal}
\usage{
createSection(title, ordinal)
}
\arguments{
\item{title}{The title of the Section}

\item{ordinal}{The ordinal of the Section}
}
\value{
A Section object
}
\description{
Create a Section object and set its title and ordinal
}
