% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createTableShell}
\alias{createTableShell}
\title{Create an empty TableShell object and set its title}
\usage{
createTableShell(title, targetCohorts, lineItems)
}
\arguments{
\item{title}{The title of the TableShell}

\item{targetCohorts}{A list of TargetCohort objects}

\item{lineItems}{A list of lineItem objects}
}
\value{
A TableShell object
}
\description{
Create an empty TableShell object and set its title
}
