% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{createTableShell}
\alias{createTableShell}
\title{Create an empty TableShell object and set its title}
\usage{
createTableShell(title)
}
\arguments{
\item{title}{The title of the TableShell}
}
\value{
A TableShell object
}
\description{
Create an empty TableShell object and set its title
}
