% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{customBreaks}
\alias{customBreaks}
\title{Function to make custom categorical breaks}
\usage{
customBreaks(name, breaks, labels)
}
\arguments{
\item{name}{the name of the categorization, as single character string}

\item{breaks}{a sequence of values to use as break points for categorization.
The sequence corresponds to the left side of the bound. For example
0, 2, 6, 12, 18 would represent age cuts of 0-1, 2-5, 6-11, 12-17 and 18+.}

\item{labels}{a sequence of character strings labelling the break points for categorization.
For example 0-1, 2-5, 6-11, 12-17 and 18+ would be labels to sequence 0, 2, 6, 12, 18.}
}
\value{
Creates a breaksStrategy object holding the labels for categorization
}
\description{
Function to make custom categorical breaks
}
