% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{indexYear}
\alias{indexYear}
\title{Create an index year characteristic}
\usage{
indexYear(breaks = NULL)
}
\arguments{
\item{breaks}{a breaksStrategy object dictating how to classify years into categories. By default this will do each year from 2000 to current day.}
}
\value{
A DemographicIndexYear Statistic class object
}
\description{
Create an index year characteristic
}
