% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{intervalRate}
\alias{intervalRate}
\title{Create a interval rate statistic}
\usage{
intervalRate(interval = c("yearly", "monthly"))
}
\arguments{
\item{interval}{a string either yearly or monthly that defines the time modifcation for the interval rate}
}
\value{
A stat object f class intervalRate
}
\description{
This statistic sums the number of occurrences of an event in a timeInterval and
divides it by the time (modified by month or year) to construct a rate per patient. This
can then be summariezed as a continuous variable
}
