% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{observedTimeToFirst}
\alias{observedTimeToFirst}
\title{Create a continuous time to stat where only occurrence during the observation period are valid}
\usage{
observedTimeToFirst()
}
\value{
A stat object continuousDistribution
}
\description{
Create a continuous time to stat where only occurrence during the observation period are valid
}
