% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{payerType}
\alias{payerType}
\title{Create an payer type char}
\usage{
payerType(breaks = NULL)
}
\arguments{
\item{breaks}{A Breaks object dictating how to classify counts into categories.
If null then this defaults to a continuous distribution}
}
\value{
A DemographicIndexYear Statistic class object
}
\description{
Create an payer type char
}
