% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{personLocation}
\alias{personLocation}
\title{Create a location char}
\usage{
personLocation(breaks)
}
\arguments{
\item{breaks}{A Breaks object dictating how to classify counts into categories.
If null then this defaults to a continuous distribution}
}
\value{
A DemographicLocation Statistic class object
}
\description{
Create a location char
}
