% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{Presence}
\alias{Presence}
\title{An R6 class to define a Presence object}
\description{
Child of Statistic. The Presence statistic is a binary metric the indicates the presence of a variable
}
\section{Super class}{
\code{\link[ClinicalCharacteristics:Statistic]{ClinicalCharacteristics::Statistic}} -> \code{Presence}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Presence-new}{\code{Presence$new()}}
\item \href{#method-Presence-clone}{\code{Presence$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ClinicalCharacteristics" data-topic="Statistic" data-id="getStatType"><a href='../../ClinicalCharacteristics/html/Statistic.html#method-Statistic-getStatType'><code>ClinicalCharacteristics::Statistic$getStatType()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Presence-new"></a>}}
\if{latex}{\out{\hypertarget{method-Presence-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Presence$new(operator, occurrences)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Presence-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Presence-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Presence$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
