% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{reviewTableShellSql}
\alias{reviewTableShellSql}
\title{Function that previews sql script used to generate results for table shell}
\usage{
reviewTableShellSql(
  tableShell,
  executionSettings,
  buildOptions = NULL,
  saveName = NULL,
  savePath = here::here()
)
}
\arguments{
\item{tableShell}{The TableShell object to used for generation}

\item{executionSettings}{The ExecutionSettings object used to generate table shell}

\item{buildOptions}{The BuildOptions object used to generate table shell}

\item{saveName}{The name of the table shell sql file}

\item{savePath}{the folder location to save the file}
}
\value{
A sql file written to a specific location
}
\description{
Function that previews sql script used to generate results for table shell
}
