% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runners.R
\name{runClinicalCharacteristics}
\alias{runClinicalCharacteristics}
\title{Function that runs the clinical characterization analysis}
\usage{
runClinicalCharacteristics(
  con,
  executionSettings,
  analysisSettings,
  outputFolder
)
}
\arguments{
\item{con}{a connection to a dbms hosting OMOP data}

\item{executionSettings}{a list of settings used to specify aspects of the connections}

\item{analysisSettings}{a list of settings specifying the details of the analysis and how the
characterization should be done}

\item{outputFolder}{a path to a folder specifying where to save the results}
}
\value{
no object return, this function will save csv files of the characterization results
}
\description{
Function that runs the clinical characterization analysis
}
