% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinChar.R
\name{runClinicalCharacteristics}
\alias{runClinicalCharacteristics}
\title{Runs the characterization and extracts data into an arrow object}
\usage{
runClinicalCharacteristics(connection, clinChar)
}
\arguments{
\item{connection}{the DatabaseConnector connection linking to the dbms with OMOP data}

\item{clinChar}{the clinChar object describing the study}
}
\value{
runs database query described in extractSettings and uploads them to the stow object
}
\description{
This runs the characterization specified by the clinChar object
}
