% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Class.R
\name{Statistic}
\alias{Statistic}
\title{An R6 class to define a Statistic object}
\description{
A Statistic is a type of metric to be used for characterization
Specific types of statistics are defined in derived classes
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Statistic-new}{\code{Statistic$new()}}
\item \href{#method-Statistic-clone}{\code{Statistic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Statistic-new"></a>}}
\if{latex}{\out{\hypertarget{method-Statistic-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Statistic$new(type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Statistic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Statistic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Statistic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
