% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6Wrappers.R
\name{timeInterval}
\alias{timeInterval}
\title{Create a single time interval}
\usage{
timeInterval(lb, rb)
}
\arguments{
\item{lb}{the left bound of the time interval}

\item{rb}{the right bound of the time interval}
}
\value{
A time interval object
}
\description{
Create a single time interval
}
